harness: split native_tools god file into 11 modules
- mod.rs: facade + re-exports (6898 -> 3834)
- common.rs: req_str, has_capability_action, etc.
- registry.rs: register_tools + schemas
- dispatch.rs: execute_tool* + capability gating
- files.rs: file CRUD + helpers (613 LoC)
- research.rs: fetch_url/web_search + SSRF (308)
- schedules.rs: upsert_schedule (156)
- goals.rs: goal CRUD (264)
- sessions.rs: read_session_messages (48)
- artifacts.rs: create/read/update/grant (443)
- tasks.rs: stub for future task extraction

All cargo check + cargo test --lib --no-run green. Avg ~560 LoC.
