+++ src/harness/native_tools/mod.rs
@@ -26,6 +26,28 @@ use crate::harness::skills::namespace::{self, NamespaceSkill};
 use crate::harness::skills::registry::ToolRegistry;
 use crate::harness::skills::render::format_active_skill_context;
 
+mod common;
+pub mod artifacts;
+pub mod tasks;
+pub mod files;
+pub mod research;
+pub mod schedules;
+pub mod goals;
+pub mod sessions;
+pub mod dispatch;
+pub use dispatch::{execute_tool, execute_tool_for_session, execute_tool_for_session_output};
+pub(crate) use artifacts::{create_artifact, read_artifact, update_artifact, get_artifact_metadata, grant_artifact, artifact_content_bytes, resolve_artifact_uri, artifact_json, default_access_expiry, parse_artifact_uri, access_expiry_from_ttl_seconds};
+pub use files::{ensure_file_read_namespace, find_file_by_path, parse_file_uri};
+pub(crate) use files::*;
+pub(crate) use research::*;
+pub(crate) use goals::*;
+pub(crate) use schedules::*;
+
+pub use goals::active_goals_context;
+mod registry;
+use common::{has_capability_action, normalize_logical_path, opt_str, opt_u64, opt_usize, req_str, require_capability, require_file_read, string_map, string_vec};
+pub use registry::{register_channel_tools, register_skill_tools, register_tools};
+
 #[path = "../tools/a2a.rs"]
 mod a2a_tools;
 #[path = "../tools/artifacts.rs"]
@@ -67,11 +89,6 @@ pub const GET_ARTIFACT_METADATA_TOOL: &str = "get_artifact_metadata";
 pub const GRANT_ARTIFACT_TOOL: &str = "grant_artifact";
 pub const FETCH_URL_TOOL: &str = "fetch_url";
 pub const WEB_SEARCH_TOOL: &str = "web_search";
-pub const SEARCH_MEMORY_TOOL: &str = "search_memory";
-pub const READ_MEMORY_TOOL: &str = "read_memory";
-pub const LIST_MEMORY_TOOL: &str = "list_memory";
-pub const CREATE_MEMORY_TOOL: &str = "create_memory";
-pub const UPDATE_MEMORY_TOOL: &str = "update_memory";
 pub const LIST_FILES_TOOL: &str = "list_files";
 pub const READ_FILE_TOOL: &str = "read_file";
 pub const GET_FILE_METADATA_TOOL: &str = "get_file_metadata";
@@ -128,2382 +145,101 @@ pub(crate) fn artifact_uris_from_message_text(text: &str) -> Vec<String> {
     a2a_tools::artifact_uris_from_message_text(text)
 }
 
-pub fn register_skill_tools(registry: &mut ToolRegistry, skills: &[NamespaceSkill]) {
-    let names = namespace::effective_skill_names(skills);
-    if !names.is_empty() {
-        registry.register_builtin(
-            ACTIVATE_SKILL_TOOL,
-            "Load the full instructions for an available namespace skill before applying it.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "name": {
-                        "type": "string",
-                        "description": "Available skill name to activate.",
-                        "enum": names
-                    }
-                },
-                "required": ["name"]
-            }),
-        );
-    }
-    registry.register_builtin(
-        DEACTIVATE_SKILL_TOOL,
-        "Stop applying an active namespace skill for the rest of this session.",
-        json!({
-            "type": "object",
-            "properties": {
-                "name": { "type": "string", "description": "Active skill name to deactivate." }
-            },
-            "required": ["name"]
-        }),
-    );
-}
 
-pub fn register_channel_tools(registry: &mut ToolRegistry) {
-    registry.register_builtin(
-        CHANNEL_PUBLISH_TOOL,
-        "Publish a public response to the channel that triggered this session. Normal assistant text remains private; use this tool for channel-visible replies.",
-        json!({
-            "type": "object",
-            "properties": {
-                "content": { "type": "string", "description": "Public channel response content." }
-            },
-            "required": ["content"]
-        }),
-    );
-    registry.register_builtin(
-        CHANNEL_SKIP_REPLY_TOOL,
-        "Mark this channel-triggered session as not needing a public channel reply.",
-        json!({
-            "type": "object",
-            "properties": {
-                "reason": { "type": "string", "description": "Optional private reason for skipping a channel reply." }
-            }
-        }),
-    );
-}
 
-pub fn register_tools(registry: &mut ToolRegistry, spec: &manifests::AgentSpec, config: &Config) {
-    artifact_tools::register(registry);
-    a2a_tools::register(registry, spec);
-    register_research_tools(registry, spec);
+// dispatch moved to dispatch.rs
+// moved read_session_messages -> sessions.rs
 
-    if !has_capability_action(spec, "schedules", "inspect")
-        && !has_capability_action(spec, "schedules", "create")
-        && !has_capability_action(spec, "schedules", "update")
-        && !has_capability_action(spec, "schedules", "delete")
-        && !has_capability_action(spec, "tasks", "inspect")
-        && !has_capability_action(spec, "tasks", "create")
-        && !has_capability_action(spec, "tasks", "update")
-        && !has_capability_action(spec, "sessions", "read:messages")
-        && !has_capability_action(spec, "memory", "inspect")
-        && !has_capability_action(spec, "memory", "read")
-        && !has_capability_action(spec, "memory", "create")
-        && !has_capability_action(spec, "memory", "update")
-        && !has_capability_action(spec, "files", "inspect")
-        && !has_capability_action(spec, "files", "read")
-        && !has_capability_action(spec, "files", "create")
-        && !has_capability_action(spec, "files", "update")
-        && !has_capability_action(spec, "files", "delete")
-        && !has_capability_action(spec, "code", "run")
-        && !has_capability_action(spec, "goals", "inspect")
-        && !has_capability_action(spec, "goals", "create")
-        && !has_capability_action(spec, "goals", "update")
-    {
-        return;
-    }
 
-    if has_capability_action(spec, "schedules", "inspect") {
-        registry.register_builtin(
-            LIST_SCHEDULES_TOOL,
-            "List schedules in a namespace. Use this to inspect existing schedule configuration and status.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Namespace to inspect. Defaults to the current agent namespace if omitted." },
-                    "agent": { "type": "string", "description": "Optional target agent filter." },
-                    "enabled": { "type": "boolean", "description": "Optional enabled-state filter." },
-                    "limit": { "type": "integer", "description": "Optional maximum number of results to return." }
-                }
-            }),
-        );
-        registry.register_builtin(
-            GET_SCHEDULE_TOOL,
-            "Get a single schedule and its runtime status.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Namespace containing the schedule. Defaults to the current agent namespace if omitted." },
-                    "name": { "type": "string", "description": "Schedule name." }
-                },
-                "required": ["name"]
-            }),
-        );
-    }
 
-    if has_capability_action(spec, "schedules", "create") {
-        registry.register_builtin(
-            CREATE_SCHEDULE_TOOL,
-            "Create a schedule directly in Talon without using talon-ops MCP.",
-            put_schedule_schema(),
-        );
-    }
-    if has_capability_action(spec, "schedules", "update") {
-        registry.register_builtin(
-            UPDATE_SCHEDULE_TOOL,
-            "Update an existing schedule directly in Talon without using talon-ops MCP.",
-            put_schedule_schema(),
-        );
-    }
-    if has_capability_action(spec, "schedules", "delete") {
-        registry.register_builtin(
-            DELETE_SCHEDULE_TOOL,
-            "Delete a schedule directly in Talon without using talon-ops MCP.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Namespace containing the schedule. Defaults to the current agent namespace if omitted." },
-                    "name": { "type": "string", "description": "Schedule name." }
-                },
-                "required": ["name"]
-            }),
-        );
-    }
 
-    if has_capability_action(spec, "sessions", "read:messages") {
-        registry.register_builtin(
-            READ_SESSION_MESSAGES_TOOL,
-            "Read text messages from a Talon session. Use this to inspect delegated child agent output by namespace, agent, and session id.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Session namespace. Defaults to current namespace." },
-                    "agent": { "type": "string", "description": "Session agent. Defaults to current agent." },
-                    "session_id": { "type": "string", "description": "Session id to inspect." },
-                    "limit": { "type": "integer", "description": "Maximum messages to return. Defaults to 20." }
-                },
-                "required": ["session_id"]
-            }),
-        );
-    }
 
-    if has_capability_action(spec, "memory", "inspect")
-        || has_capability_action(spec, "memory", "read")
-    {
-        registry.register_builtin(
-            SEARCH_MEMORY_TOOL,
-            "Search durable workspace memory Files. Memory is stored as namespace File resources with purpose=MEMORY and indexPolicy=RETRIEVAL.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Memory namespace. Defaults to current namespace." },
-                    "query": { "type": "string", "description": "Keyword query to match against memory path and text content." },
-                    "prefix": { "type": "string", "description": "Optional logical path prefix such as /memory/playbooks." },
-                    "limit": { "type": "integer", "description": "Maximum results. Defaults to 10." }
-                },
-                "required": ["query"]
-            }),
-        );
-        registry.register_builtin(
-            READ_MEMORY_TOOL,
-            "Read one durable workspace memory File by logical path.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Memory namespace. Defaults to current namespace." },
-                    "path": { "type": "string", "description": "Logical memory path, for example /memory/playbooks/aeo-prompt-strategy.md." }
-                },
-                "required": ["path"]
-            }),
-        );
-        registry.register_builtin(
-            LIST_MEMORY_TOOL,
-            "List durable workspace memory Files by optional logical path prefix.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Memory namespace. Defaults to current namespace." },
-                    "prefix": { "type": "string", "description": "Optional logical path prefix. Defaults to /memory." },
-                    "limit": { "type": "integer", "description": "Maximum results. Defaults to 50." }
-                }
-            }),
-        );
-    }
+// moved list_files_tool -> files.rs
 
-    if has_capability_action(spec, "memory", "create") {
-        registry.register_builtin(
-            CREATE_MEMORY_TOOL,
-            "Create a durable workspace memory File with purpose=MEMORY and indexPolicy=RETRIEVAL.",
-            memory_write_schema(),
-        );
-    }
-    if has_capability_action(spec, "memory", "update") {
-        registry.register_builtin(
-            UPDATE_MEMORY_TOOL,
-            "Update a durable workspace memory File. Updates write a new immutable CAS object and advance File.status.objectRef.",
-            memory_write_schema(),
-        );
-    }
+// moved read_file_tool -> files.rs
 
-    if has_capability_action(spec, "files", "inspect")
-        || has_capability_action(spec, "files", "read")
-    {
-        registry.register_builtin(
-            LIST_FILES_TOOL,
-            "List namespace File resources by optional path prefix, purpose, and index policy.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "File namespace. Defaults to current namespace." },
-                    "prefix": { "type": "string", "description": "Optional logical path prefix such as /content/pages." },
-                    "purpose": { "type": "string", "description": "Optional purpose: ARTIFACT, MEMORY, or SKILL." },
-                    "index_policy": { "type": "string", "description": "Optional index policy: NONE, SEARCH, or RETRIEVAL." },
-                    "limit": { "type": "integer", "description": "Maximum results. Defaults to 50." }
-                }
-            }),
-        );
-        registry.register_builtin(
-            READ_FILE_TOOL,
-            "Read a namespace File by file:// URI or logical path.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "uri": { "type": "string", "description": "Optional file://<namespace>/<path> URI returned by Talon or Conic." },
-                    "namespace": { "type": "string", "description": "File namespace. Defaults to current namespace." },
-                    "path": { "type": "string", "description": "Logical File path, for example /content/pages/<id>/content.md." }
-                }
-            }),
-        );
-        registry.register_builtin(
-            GET_FILE_METADATA_TOOL,
-            "Get File metadata by file:// URI or logical path without reading content.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "uri": { "type": "string", "description": "Optional file://<namespace>/<path> URI returned by Talon or Conic." },
-                    "namespace": { "type": "string", "description": "File namespace. Defaults to current namespace." },
-                    "path": { "type": "string", "description": "Logical File path." }
-                }
-            }),
-        );
-    }
+// moved get_file_metadata_tool -> files.rs
 
-    if has_capability_action(spec, "files", "create") {
-        registry.register_builtin(
-            CREATE_FILE_TOOL,
-            "Create a namespace File. Defaults to purpose=ARTIFACT, index_policy=SEARCH, and retention=RETAINED.",
-            file_write_schema(),
-        );
-    }
-    if has_capability_action(spec, "files", "update") {
-        registry.register_builtin(
-            UPDATE_FILE_TOOL,
-            "Update an existing namespace File by file:// URI or logical path.",
-            file_write_schema(),
-        );
-    }
-    if has_capability_action(spec, "files", "delete") {
-        registry.register_builtin(
-            DELETE_FILE_TOOL,
-            "Delete a namespace File resource by file:// URI or logical path.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "uri": { "type": "string", "description": "Optional file://<namespace>/<path> URI returned by Talon or Conic." },
-                    "namespace": { "type": "string", "description": "File namespace. Defaults to current namespace." },
-                    "path": { "type": "string", "description": "Logical File path." }
-                }
-            }),
-        );
-    }
+// moved create_file_tool -> files.rs
 
-    code_tools::register(registry, spec, config);
+// moved update_file_tool -> files.rs
 
-    task_tools::register(registry, spec);
+// moved delete_file_tool -> files.rs
 
-    if has_capability_action(spec, "goals", "inspect") {
-        registry.register_builtin(
-            LIST_GOALS_TOOL,
-            "List Talon Goals for one session.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Namespace to inspect. Defaults to the current namespace." },
-                    "agent": { "type": "string", "description": "Owning agent. Defaults to the current agent." },
-                    "session_id": { "type": "string", "description": "Owning session id. Defaults to the current session." },
-                    "status_group": { "type": "string", "description": "Optional group: active or terminal." },
-                    "phase": { "type": "string", "description": "Optional phase such as RUNNING, NEEDS_REVIEW, SUCCEEDED, FAILED, BLOCKED, or CANCELED." },
-                    "limit": { "type": "integer", "description": "Optional maximum number of goals to return." }
-                }
-            }),
-        );
-        registry.register_builtin(
-            GET_GOAL_TOOL,
-            "Get one Talon Goal by id.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Goal namespace. Defaults to the current namespace." },
-                    "agent": { "type": "string", "description": "Owning agent. Defaults to the current agent." },
-                    "session_id": { "type": "string", "description": "Owning session id. Defaults to the current session." },
-                    "goal_id": { "type": "string", "description": "Goal id." }
-                },
-                "required": ["goal_id"]
-            }),
-        );
-    }
 
-    if has_capability_action(spec, "goals", "create") {
-        registry.register_builtin(
-            CREATE_GOAL_TOOL,
-            "Create a session-scoped Talon Goal that tracks a durable objective and success criteria.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "objective": { "type": "string", "description": "Durable objective the agent should keep in view." },
-                    "success_criteria": { "type": "array", "items": { "type": "string" }, "description": "Concrete completion criteria." },
-                    "max_iterations": { "type": "integer", "description": "Optional maximum iteration count." },
-                    "progress_summary": { "type": "string", "description": "Optional initial progress summary." },
-                    "labels": { "type": "object", "additionalProperties": { "type": "string" } },
-                    "metadata": { "type": "object", "additionalProperties": { "type": "string" } }
-                },
-                "required": ["objective"]
-            }),
-        );
-    }
+// moved list_files_by_filter -> files.rs
 
-    if has_capability_action(spec, "goals", "update") {
-        registry.register_builtin(
-            UPDATE_GOAL_TOOL,
-            "Update Goal phase, progress, iteration, or blocked reason.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Goal namespace. Defaults to current namespace." },
-                    "agent": { "type": "string", "description": "Owning agent. Defaults to current agent." },
-                    "session_id": { "type": "string", "description": "Owning session id. Defaults to current session." },
-                    "goal_id": { "type": "string", "description": "Goal id." },
-                    "phase": { "type": "string", "description": "RUNNING, PAUSED, NEEDS_REVIEW, SUCCEEDED, FAILED, BLOCKED, CANCELED, or EXPIRED." },
-                    "progress_summary": { "type": "string", "description": "Concise current state." },
-                    "iteration": { "type": "integer", "description": "Current iteration number." },
-                    "blocked_reason": { "type": "string", "description": "Reason the Goal is blocked." }
-                },
-                "required": ["goal_id"]
-            }),
-        );
-        registry.register_builtin(
-            COMPLETE_GOAL_TOOL,
-            "Mark a Goal as SUCCEEDED with an optional final progress summary.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Goal namespace. Defaults to current namespace." },
-                    "agent": { "type": "string", "description": "Owning agent. Defaults to current agent." },
-                    "session_id": { "type": "string", "description": "Owning session id. Defaults to current session." },
-                    "goal_id": { "type": "string", "description": "Goal id." },
-                    "progress_summary": { "type": "string", "description": "Final result summary." }
-                },
-                "required": ["goal_id"]
-            }),
-        );
-        registry.register_builtin(
-            BLOCK_GOAL_TOOL,
-            "Mark a Goal as BLOCKED with the reason no meaningful progress can continue.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "namespace": { "type": "string", "description": "Goal namespace. Defaults to current namespace." },
-                    "agent": { "type": "string", "description": "Owning agent. Defaults to current agent." },
-                    "session_id": { "type": "string", "description": "Owning session id. Defaults to current session." },
-                    "goal_id": { "type": "string", "description": "Goal id." },
-                    "blocked_reason": { "type": "string", "description": "Concrete blocker." },
-                    "progress_summary": { "type": "string", "description": "Optional progress summary." }
-                },
-                "required": ["goal_id", "blocked_reason"]
-            }),
-        );
-    }
-}
 
-fn register_research_tools(registry: &mut ToolRegistry, spec: &manifests::AgentSpec) {
-    if has_capability_action(spec, "research", "fetch_url") {
-        registry.register_builtin(
-            FETCH_URL_TOOL,
-            "Fetch a supplied HTTP(S) URL and return title, final URL, status, and compact visible text for source-grounded research.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "url": { "type": "string", "description": "HTTP(S) URL to fetch." },
-                    "max_chars": {
-                        "type": "integer",
-                        "description": "Maximum visible text characters to return. Defaults to 12000."
-                    }
-                },
-                "required": ["url"]
-            }),
-        );
-    }
-
-    if has_capability_action(spec, "research", "web_search") {
-        registry.register_builtin(
-            WEB_SEARCH_TOOL,
-            "Search the public web for source candidates. Use returned URLs with fetch_url before citing claims.",
-            json!({
-                "type": "object",
-                "properties": {
-                    "query": { "type": "string", "description": "Search query." },
-                    "limit": {
-                        "type": "integer",
-                        "description": "Maximum number of search results. Defaults to 5."
-                    }
-                },
-                "required": ["query"]
-            }),
-        );
-    }
-}
-
-fn put_schedule_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "namespace": { "type": "string", "description": "Target namespace. Defaults to the current agent namespace if omitted." },
-            "name": { "type": "string", "description": "Schedule name." },
-            "labels": {
-                "type": "object",
-                "description": "Optional schedule labels.",
-                "additionalProperties": { "type": "string" }
-            },
-            "kind": { "type": "string", "description": "Schedule kind: at, every, or cron." },
-            "cron": { "type": "string", "description": "Cron expression for cron schedules." },
-            "interval_seconds": { "type": "integer", "description": "Interval in seconds for every schedules." },
-            "run_at": { "type": "string", "description": "RFC3339 timestamp for at schedules." },
-            "timezone": { "type": "string", "description": "Optional timezone." },
-            "agent": { "type": "string", "description": "Target agent. Defaults to the current agent if omitted." },
-            "session_mode": { "type": "string", "description": "Session mode: new or reuse." },
-            "session_id": { "type": "string", "description": "Session id to reuse when session_mode is reuse." },
-            "input_message": { "type": "string", "description": "Message the schedule should send when it runs." },
-            "enabled": { "type": "boolean", "description": "Whether the schedule is enabled." }
-        },
-        "required": ["name", "kind", "input_message"]
-    })
-}
-
-fn memory_write_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "namespace": { "type": "string", "description": "Memory namespace. Defaults to current namespace." },
-            "path": { "type": "string", "description": "Logical memory path, for example /memory/research/context.md." },
-            "content": { "type": "string", "description": "Markdown or text content to store." },
-            "media_type": { "type": "string", "description": "Media type. Defaults to text/markdown." }
-        },
-        "required": ["path", "content"]
-    })
-}
-
-fn file_write_schema() -> Value {
-    json!({
-        "type": "object",
-        "properties": {
-            "uri": { "type": "string", "description": "Optional file://<namespace>/<path> URI for updates." },
-            "namespace": { "type": "string", "description": "File namespace. Defaults to current namespace." },
-            "path": { "type": "string", "description": "Logical File path, for example /content/pages/<id>/content.md." },
-            "content": { "type": "string", "description": "Markdown, HTML, or text content to store." },
-            "media_type": { "type": "string", "description": "Media type. Defaults to text/markdown." },
-            "purpose": { "type": "string", "description": "File purpose: ARTIFACT, MEMORY, or SKILL. Defaults to ARTIFACT." },
-            "index_policy": { "type": "string", "description": "Index policy: NONE, SEARCH, or RETRIEVAL. Defaults to SEARCH." },
-            "retention": { "type": "string", "description": "Retention policy: RETAINED. Defaults to RETAINED." }
-        },
-        "required": ["content"]
-    })
-}
-
-pub async fn execute_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<String>> {
-    execute_tool_for_session(
-        cp,
-        current_namespace,
-        current_agent,
-        "",
-        spec,
-        name,
-        args,
-        config,
-    )
-    .await
-}
-
-pub async fn execute_tool_for_session(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<String>> {
-    Ok(execute_tool_for_session_output(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-        config,
-    )
-    .await?
-    .map(|output| output.summary()))
-}
-
-pub async fn execute_tool_for_session_output(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    spec: &manifests::AgentSpec,
-    name: &str,
-    args: &Value,
-    config: &Config,
-) -> Result<Option<ToolOutput>> {
-    if let Some((capability, action)) = global_capability_for_tool(name) {
-        require_global_capability(config, capability, action)?;
-    }
-
-    if let Some(result) = artifact_tools::execute_output(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(result));
-    }
-    if let Some(result) = a2a_tools::execute(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(ToolOutput::text(result)));
-    }
-    if let Some(result) = task_tools::execute(
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    )
-    .await?
-    {
-        return Ok(Some(ToolOutput::text(result)));
-    }
-    // Keep the Monty execution future off the Tokio worker stack.  Its
-    // subprocess/lifecycle state is substantially larger than the generic
-    // native-tool dispatch future, and this branch is also evaluated for
-    // unrelated tools.
-    if let Some(result) = Box::pin(code_tools::execute(
-        config,
-        cp,
-        current_namespace,
-        current_agent,
-        current_session,
-        spec,
-        name,
-        args,
-    ))
-    .await?
-    {
-        return Ok(Some(result));
-    }
-
-    match name {
-        READ_SESSION_MESSAGES_TOOL => {
-            require_capability(spec, "sessions", "read:messages")?;
-            read_session_messages(cp, current_namespace, current_agent, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        SEARCH_MEMORY_TOOL => {
-            require_memory_read(spec)?;
-            search_memory(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        READ_MEMORY_TOOL => {
-            require_memory_read(spec)?;
-            read_memory(cp, current_namespace, args).await.map(Some)
-        }
-        LIST_MEMORY_TOOL => {
-            require_memory_read(spec)?;
-            list_memory(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        CREATE_MEMORY_TOOL => {
-            require_capability(spec, "memory", "create")?;
-            put_memory(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        UPDATE_MEMORY_TOOL => {
-            require_capability(spec, "memory", "update")?;
-            put_memory(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        LIST_FILES_TOOL => {
-            require_file_read(spec)?;
-            list_files_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        READ_FILE_TOOL => {
-            require_file_read(spec)?;
-            read_file_tool(cp, current_namespace, args).await.map(Some)
-        }
-        GET_FILE_METADATA_TOOL => {
-            require_file_read(spec)?;
-            get_file_metadata_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        CREATE_FILE_TOOL => {
-            require_capability(spec, "files", "create")?;
-            create_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        UPDATE_FILE_TOOL => {
-            require_capability(spec, "files", "update")?;
-            update_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        DELETE_FILE_TOOL => {
-            require_capability(spec, "files", "delete")?;
-            delete_file_tool(cp, current_namespace, args)
-                .await
-                .map(ToolOutput::text)
-                .map(Some)
-        }
-        FETCH_URL_TOOL => {
-            require_capability(spec, "research", "fetch_url")?;
-            fetch_url(args).await.map(ToolOutput::text).map(Some)
-        }
-        WEB_SEARCH_TOOL => {
-            require_capability(spec, "research", "web_search")?;
-            web_search(args).await.map(ToolOutput::text).map(Some)
-        }
-        ACTIVATE_SKILL_TOOL => {
-            let skill_name = req_str(args, "name")?;
-            let skills = namespace::load_available_skills(cp, current_namespace).await?;
-            let skill = namespace::find_effective_skill(&skills, skill_name)
-                .ok_or_else(|| anyhow!("skill '{}' is not available", skill_name))?;
-            let instructions = namespace::load_skill_instructions(cp, skill).await?;
-            if current_session.is_empty() {
-                return Ok(Some(ToolOutput::text(format_active_skill_context(&[(
-                    skill.clone(),
-                    instructions,
-                )]))));
-            }
-            crate::harness::sessions::activate_skill(
-                cp.kv.as_ref(),
-                current_namespace,
-                current_agent,
-                current_session,
-                skill_name,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(format!(
-                "Activated skill '{}'. Its workflow guidance is now active for this session.",
-                skill_name
-            ))))
-        }
-        DEACTIVATE_SKILL_TOOL => {
-            let skill_name = req_str(args, "name")?;
-            if current_session.is_empty() {
-                return Err(anyhow!("deactivate_skill requires a session"));
-            }
-            let removed = crate::harness::sessions::deactivate_skill(
-                cp.kv.as_ref(),
-                current_namespace,
-                current_agent,
-                current_session,
-                skill_name,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(if removed {
-                format!("Deactivated skill '{}'.", skill_name)
-            } else {
-                format!("Skill '{}' was not active.", skill_name)
-            })))
-        }
-        CHANNEL_PUBLISH_TOOL => {
-            let content = req_str(args, "content")?;
-            let message = crate::gateway::rpc::channels::publish_channel_message_from_session(
-                cp,
-                current_namespace,
-                current_agent,
-                current_session,
-                content,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "published": true,
-                    "messageId": message.id,
-                    "channel": message.channel
-                }),
-            )?)))
-        }
-        CHANNEL_SKIP_REPLY_TOOL => {
-            let reason = opt_str(args, "reason").unwrap_or("");
-            crate::gateway::rpc::channels::skip_channel_reply_from_session(
-                cp,
-                current_namespace,
-                current_agent,
-                current_session,
-                reason,
-            )
-            .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "published": false,
-                    "skipped": true
-                }),
-            )?)))
-        }
-        LIST_SCHEDULES_TOOL => {
-            require_capability(spec, "schedules", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let agent = opt_str(args, "agent");
-            let enabled = args.get("enabled").and_then(Value::as_bool);
-            let limit = args.get("limit").and_then(Value::as_u64).unwrap_or(100) as usize;
-            let entries = cp
-                .kv
-                .list_entries(&keys::schedule_prefix(namespace), None)
-                .await?;
-            let mut schedules = Vec::new();
-            for (_key, value) in entries {
-                let schedule = resources_proto::Schedule::decode(value.as_slice())?;
-                let spec_model = schedule.spec.as_ref();
-                let matches_agent = agent
-                    .map(|target| {
-                        spec_model
-                            .and_then(|current| current.target.as_ref())
-                            .map(|target_model| target_model.agent == target)
-                            .unwrap_or(false)
-                    })
-                    .unwrap_or(true);
-                let matches_enabled = enabled
-                    .map(|value| {
-                        spec_model
-                            .map(|current| current.enabled == value)
-                            .unwrap_or(false)
-                    })
-                    .unwrap_or(true);
-                if matches_agent && matches_enabled {
-                    schedules.push(schedule_json(&schedule));
-                }
-                if schedules.len() >= limit {
-                    break;
-                }
-            }
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({ "schedules": schedules }),
-            )?)))
-        }
-        GET_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let schedule = cp
-                .kv
-                .get_msg::<resources_proto::Schedule>(&keys::schedule(namespace, schedule_name))
-                .await?
-                .ok_or_else(|| anyhow!("schedule '{}' not found", schedule_name))?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule)
-                }),
-            )?)))
-        }
-        CREATE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "create")?;
-            let schedule =
-                upsert_schedule(cp, current_namespace, current_agent, args, None).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule),
-                    "backendArmed": schedule.status.as_ref().map(|status| status.backend_armed).unwrap_or(false)
-                }),
-            )?)))
-        }
-        UPDATE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "update")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let existing = cp
-                .kv
-                .get_msg::<resources_proto::Schedule>(&keys::schedule(namespace, schedule_name))
-                .await?
-                .ok_or_else(|| anyhow!("schedule '{}' not found", schedule_name))?;
-            let schedule =
-                upsert_schedule(cp, current_namespace, current_agent, args, Some(existing)).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "schedule": schedule_json(&schedule),
-                    "backendArmed": schedule.status.as_ref().map(|status| status.backend_armed).unwrap_or(false)
-                }),
-            )?)))
-        }
-        DELETE_SCHEDULE_TOOL => {
-            require_capability(spec, "schedules", "delete")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let schedule_name = req_str(args, "name")?;
-            let key = keys::schedule(namespace, schedule_name);
-            if let Some(schedule) = cp.kv.get_msg::<resources_proto::Schedule>(&key).await? {
-                if let Some(handle) = schedule.status.and_then(|status| status.backend_handle) {
-                    if let Err(error) = cp.scheduler.cancel(&handle).await {
-                        tracing::warn!(handle = %handle, error = %error, "failed to cancel schedule handle");
-                    }
-                }
-            }
-            cp.kv.delete(&key).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({ "success": true }),
-            )?)))
-        }
-        LIST_GOALS_TOOL => {
-            require_capability(spec, "goals", "inspect")?;
-            let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-            let agent = opt_str(args, "agent").unwrap_or(current_agent);
-            let session_id = opt_str(args, "session_id").unwrap_or(current_session);
-            let status_group = opt_str(args, "status_group");
-            let phase = opt_str(args, "phase");
-            let limit = args.get("limit").and_then(Value::as_u64).unwrap_or(100) as usize;
-            let goals = list_goals(cp, namespace, agent, session_id, status_group, phase, limit)
-                .await?
-                .into_iter()
-                .map(|goal| goal_json(&goal))
-                .collect::<Vec<_>>();
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goals": goals
-                }),
-            )?)))
-        }
-        GET_GOAL_TOOL => {
-            require_capability(spec, "goals", "inspect")?;
-            let goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        CREATE_GOAL_TOOL => {
-            require_capability(spec, "goals", "create")?;
-            let goal =
-                create_goal(cp, current_namespace, current_agent, current_session, args).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        UPDATE_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            update_goal_from_args(&mut goal, args)?;
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        COMPLETE_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            let now = chrono::Utc::now().timestamp_micros();
-            goal.phase = crate::gateway::rpc::data_proto::GoalPhase::Succeeded as i32;
-            goal.updated_at = now;
-            goal.completed_at = now;
-            if let Some(summary) = opt_str(args, "progress_summary") {
-                goal.progress_summary = summary.to_string();
-            }
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        BLOCK_GOAL_TOOL => {
-            require_capability(spec, "goals", "update")?;
-            let mut goal =
-                get_goal_from_args(cp, current_namespace, current_agent, current_session, args)
-                    .await?;
-            let now = chrono::Utc::now().timestamp_micros();
-            goal.phase = crate::gateway::rpc::data_proto::GoalPhase::Blocked as i32;
-            goal.updated_at = now;
-            goal.blocked_reason = req_str(args, "blocked_reason")?.to_string();
-            if let Some(summary) = opt_str(args, "progress_summary") {
-                goal.progress_summary = summary.to_string();
-            }
-            upsert_goal(cp, goal.clone()).await?;
-            Ok(Some(ToolOutput::text(serde_json::to_string_pretty(
-                &json!({
-                    "goal": goal_json(&goal)
-                }),
-            )?)))
-        }
-        _ => Ok(None),
-    }
-}
-
-fn require_global_capability(config: &Config, capability: &str, action: &str) -> Result<()> {
-    if global_capability_allowed(config, capability, action) {
-        Ok(())
-    } else {
-        Err(anyhow!(
-            "capability '{}:{}' is disabled by deployment configuration",
-            capability,
-            action
-        ))
-    }
-}
-
-fn global_capability_for_tool(name: &str) -> Option<(&'static str, &'static str)> {
-    match name {
-        READ_SESSION_MESSAGES_TOOL => Some(("sessions", "read:messages")),
-        SEARCH_MEMORY_TOOL | READ_MEMORY_TOOL | LIST_MEMORY_TOOL => Some(("memory", "read")),
-        CREATE_MEMORY_TOOL => Some(("memory", "create")),
-        UPDATE_MEMORY_TOOL => Some(("memory", "update")),
-        LIST_FILES_TOOL | READ_FILE_TOOL | GET_FILE_METADATA_TOOL => Some(("files", "read")),
-        CREATE_FILE_TOOL => Some(("files", "create")),
-        UPDATE_FILE_TOOL => Some(("files", "update")),
-        DELETE_FILE_TOOL => Some(("files", "delete")),
-        FETCH_URL_TOOL => Some(("research", "fetch_url")),
-        WEB_SEARCH_TOOL => Some(("research", "web_search")),
-        LIST_SCHEDULES_TOOL | GET_SCHEDULE_TOOL => Some(("schedules", "inspect")),
-        CREATE_SCHEDULE_TOOL => Some(("schedules", "create")),
-        UPDATE_SCHEDULE_TOOL => Some(("schedules", "update")),
-        DELETE_SCHEDULE_TOOL => Some(("schedules", "delete")),
-        GET_GOAL_TOOL | LIST_GOALS_TOOL => Some(("goals", "inspect")),
-        CREATE_GOAL_TOOL => Some(("goals", "create")),
-        UPDATE_GOAL_TOOL | COMPLETE_GOAL_TOOL | BLOCK_GOAL_TOOL => Some(("goals", "update")),
-        _ => None,
-    }
-}
-
-async fn read_session_messages(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    args: &Value,
-) -> Result<String> {
-    let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-    let agent = opt_str(args, "agent").unwrap_or(current_agent);
-    let session_id = req_str(args, "session_id")?;
-    let limit = opt_usize(args, "limit").unwrap_or(20).clamp(1, 100);
-    let mut entries = cp
-        .kv
-        .list_entries(
-            &keys::session_message_prefix(namespace, agent, session_id),
-            Some(ListOptions::desc().limit(limit)),
-        )
-        .await?;
-    let mut messages = Vec::new();
-    entries.reverse();
-    for (_, bytes) in entries {
-        let message = data_proto::SessionMessage::decode(bytes.as_slice())?;
-        let role = data_proto::MessageRole::try_from(message.role)
-            .map(|role| format!("{role:?}"))
-            .unwrap_or_else(|_| message.role.to_string());
-        let text = message
-            .parts
-            .iter()
-            .filter(|part| part.part_type == data_proto::SessionMessagePartType::Text as i32)
-            .map(|part| part.content.as_str())
-            .collect::<String>();
-        messages.push(json!({
-            "id": message.id,
-            "role": role,
-            "text": text,
-            "createdAt": message.created_at,
-            "labels": message.labels,
-        }));
-    }
-    Ok(serde_json::to_string_pretty(&json!({
-        "namespace": namespace,
-        "agent": agent,
-        "sessionId": session_id,
-        "messages": messages,
-    }))?)
-}
-
-async fn search_memory(cp: &ControlPlane, current_namespace: &str, args: &Value) -> Result<String> {
-    let namespace = namespace_arg(current_namespace, args);
-    let query = req_str(args, "query")?.to_ascii_lowercase();
-    let prefix = opt_str(args, "prefix")
-        .map(normalize_logical_path)
-        .transpose()?
-        .unwrap_or_else(|| "/memory".to_string());
-    let limit = opt_usize(args, "limit").unwrap_or(10).clamp(1, 50);
-    let mut results = Vec::new();
-    for file in list_memory_files(cp, &namespace, &prefix).await? {
-        let Some(spec) = file.spec.as_ref() else {
-            continue;
-        };
-        let content = read_file_content(cp, &file).await.unwrap_or_default();
-        let haystack = format!("{}\n{}", spec.path, content).to_ascii_lowercase();
-        if haystack.contains(&query) {
-            results.push(json!({
-                "namespace": namespace,
-                "name": file_name_from_file(&file),
-                "path": spec.path,
-                "mediaType": spec.media_type,
-                "excerpt": memory_excerpt(&content, &query),
-            }));
-            if results.len() >= limit {
-                break;
-            }
-        }
-    }
-    Ok(serde_json::to_string_pretty(
-        &json!({ "results": results }),
-    )?)
-}
-
-async fn read_memory(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<ToolOutput> {
-    let namespace = namespace_arg(current_namespace, args);
-    let path = normalize_logical_path(req_str(args, "path")?)?;
-    let file = find_memory_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("memory file '{}' not found", path))?;
-    read_file_output(cp, &file).await
-}
-
-async fn list_memory(cp: &ControlPlane, current_namespace: &str, args: &Value) -> Result<String> {
-    let namespace = namespace_arg(current_namespace, args);
-    let prefix = opt_str(args, "prefix")
-        .map(normalize_logical_path)
-        .transpose()?
-        .unwrap_or_else(|| "/memory".to_string());
-    let limit = opt_usize(args, "limit").unwrap_or(50).clamp(1, 100);
-    let files = list_memory_files(cp, &namespace, &prefix).await?;
-    let entries = files
-        .into_iter()
-        .take(limit)
-        .filter_map(|file| {
-            let spec = file.spec.as_ref()?;
-            let object = file
-                .status
-                .as_ref()
-                .and_then(|status| status.object_ref.as_ref());
-            Some(json!({
-                "namespace": namespace,
-                "name": file_name_from_file(&file),
-                "path": spec.path,
-                "mediaType": spec.media_type,
-                "sizeBytes": object.map(|object| object.size_bytes).unwrap_or_default(),
-                "sha256": object.map(|object| object.sha256.as_str()).unwrap_or_default(),
-            }))
-        })
-        .collect::<Vec<_>>();
-    Ok(serde_json::to_string_pretty(
-        &json!({ "entries": entries }),
-    )?)
-}
-
-async fn put_memory(cp: &ControlPlane, current_namespace: &str, args: &Value) -> Result<String> {
-    let namespace = namespace_arg(current_namespace, args);
-    let path = normalize_memory_path(req_str(args, "path")?)?;
-    let content = req_str(args, "content")?;
-    let media_type = opt_str(args, "media_type").unwrap_or("text/markdown");
-    let file = upsert_memory_file(cp, &namespace, &path, media_type, content.as_bytes()).await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "file": memory_file_json(&file),
-    }))?)
-}
-
-async fn list_files_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let namespace = namespace_arg(current_namespace, args);
-    let prefix = opt_str(args, "prefix")
-        .map(normalize_logical_path)
-        .transpose()?;
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?;
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?;
-    let limit = opt_usize(args, "limit").unwrap_or(50).clamp(1, 100);
-    let entries = list_files_by_filter(
-        cp,
-        &namespace,
-        prefix.as_deref().unwrap_or(""),
-        purpose,
-        index_policy,
-    )
-    .await?
-    .into_iter()
-    .take(limit)
-    .map(|file| file_json(&file, false))
-    .collect::<Vec<_>>();
-    Ok(serde_json::to_string_pretty(
-        &json!({ "entries": entries }),
-    )?)
-}
-
-async fn read_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<ToolOutput> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    ensure_file_read_namespace(current_namespace, &namespace)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    read_file_output(cp, &file).await
-}
-
-async fn get_file_metadata_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    ensure_file_read_namespace(current_namespace, &namespace)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    Ok(serde_json::to_string_pretty(&file_json(&file, false))?)
-}
-
-async fn create_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    if find_file_by_path(cp, &namespace, &path).await?.is_some() {
-        return Err(anyhow!("File '{}' already exists", path));
-    }
-    let content = req_str(args, "content")?;
-    let media_type = opt_str(args, "media_type").unwrap_or("text/markdown");
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?
-        .unwrap_or(resources_proto::FilePurpose::Artifact as i32);
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?
-        .unwrap_or(resources_proto::FileIndexPolicy::Search as i32);
-    let retention = opt_str(args, "retention")
-        .map(parse_file_retention)
-        .transpose()?
-        .unwrap_or(resources_proto::FileRetention::Retained as i32);
-    let file = upsert_file(
-        cp,
-        &namespace,
-        None,
-        &path,
-        media_type,
-        purpose,
-        index_policy,
-        retention,
-        content.as_bytes(),
-    )
-    .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "file": file_json(&file, false),
-    }))?)
-}
-
-async fn update_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    let existing = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    let content = req_str(args, "content")?;
-    let existing_spec = existing.spec.as_ref();
-    let media_type = opt_str(args, "media_type")
-        .or_else(|| existing_spec.map(|spec| spec.media_type.as_str()))
-        .unwrap_or("text/markdown")
-        .to_string();
-    let purpose = opt_str(args, "purpose")
-        .map(parse_file_purpose)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.purpose))
-        .unwrap_or(resources_proto::FilePurpose::Artifact as i32);
-    let index_policy = opt_str(args, "index_policy")
-        .map(parse_file_index_policy)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.index_policy))
-        .unwrap_or(resources_proto::FileIndexPolicy::Search as i32);
-    let retention = opt_str(args, "retention")
-        .map(parse_file_retention)
-        .transpose()?
-        .or_else(|| existing_spec.map(|spec| spec.retention))
-        .unwrap_or(resources_proto::FileRetention::Retained as i32);
-    let file = upsert_file(
-        cp,
-        &namespace,
-        Some(existing),
-        &path,
-        &media_type,
-        purpose,
-        index_policy,
-        retention,
-        content.as_bytes(),
-    )
-    .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "file": file_json(&file, false),
-    }))?)
-}
-
-async fn delete_file_tool(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    args: &Value,
-) -> Result<String> {
-    let (namespace, path) = file_location_from_args(current_namespace, args)?;
-    let file = find_file_by_path(cp, &namespace, &path)
-        .await?
-        .ok_or_else(|| anyhow!("File '{}' not found", path))?;
-    let name = file_name_from_file(&file);
-    let latest_key = crate::control::cas::latest_file_object_key(&namespace, &path);
-    let object_key = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref())
-        .map(|object| object.key.clone());
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let deleted = store.delete(&namespace, "File", &name).await?;
-    if deleted {
-        let cas = crate::control::cas::CasStore::new(cp.objects.clone());
-        if let Err(error) = cas.delete_object(&latest_key).await {
-            tracing::warn!(
-                error = %error,
-                object_key = %latest_key,
-                "failed to delete latest File object from native tool"
-            );
-        }
-        if let Some(object_key) = object_key {
-            if let Err(error) = cas.delete_object(&object_key).await {
-                tracing::warn!(
-                    error = %error,
-                    object_key = %object_key,
-                    "failed to delete File CAS object from native tool"
-                );
-            }
-        }
-    }
-    Ok(serde_json::to_string_pretty(&json!({
-        "deleted": deleted,
-        "namespace": namespace,
-        "path": path,
-    }))?)
-}
-
-async fn list_memory_files(
-    cp: &ControlPlane,
-    namespace: &str,
-    prefix: &str,
-) -> Result<Vec<resources_proto::File>> {
-    list_files_by_filter(
-        cp,
-        namespace,
-        prefix,
-        Some(resources_proto::FilePurpose::Memory as i32),
-        Some(resources_proto::FileIndexPolicy::Retrieval as i32),
-    )
-    .await
-}
-
-async fn list_files_by_filter(
-    cp: &ControlPlane,
-    namespace: &str,
-    prefix: &str,
-    purpose: Option<i32>,
-    index_policy: Option<i32>,
-) -> Result<Vec<resources_proto::File>> {
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let mut files = Vec::new();
-    for resource in store.list(namespace, Some("File")).await? {
-        let Some(file) = file_from_resource(resource) else {
-            continue;
-        };
-        let Some(spec) = file.spec.as_ref() else {
-            continue;
-        };
-        if purpose.is_some_and(|purpose| spec.purpose != purpose) {
-            continue;
-        }
-        if index_policy.is_some_and(|index_policy| spec.index_policy != index_policy) {
-            continue;
-        }
-        if !prefix.is_empty() && !spec.path.starts_with(prefix) {
-            continue;
-        }
-        files.push(file);
-    }
-    files.sort_by(|left, right| {
-        left.spec
-            .as_ref()
-            .map(|spec| spec.path.as_str())
-            .cmp(&right.spec.as_ref().map(|spec| spec.path.as_str()))
-    });
-    Ok(files)
-}
-
-async fn find_memory_file_by_path(
-    cp: &ControlPlane,
-    namespace: &str,
-    path: &str,
-) -> Result<Option<resources_proto::File>> {
-    Ok(list_memory_files(cp, namespace, path)
-        .await?
-        .into_iter()
-        .find(|file| file.spec.as_ref().map(|spec| spec.path.as_str()) == Some(path)))
-}
-
-async fn find_file_by_path(
-    cp: &ControlPlane,
-    namespace: &str,
-    path: &str,
-) -> Result<Option<resources_proto::File>> {
-    Ok(list_files_by_filter(cp, namespace, path, None, None)
-        .await?
-        .into_iter()
-        .find(|file| file.spec.as_ref().map(|spec| spec.path.as_str()) == Some(path)))
-}
+// moved find_file_by_path -> files.rs
 
 struct ReadFileObject {
     object: crate::control::object_store::StoredObject,
     object_ref: data_proto::ObjectRef,
 }
 
-async fn read_file_content(cp: &ControlPlane, file: &resources_proto::File) -> Result<String> {
-    let read_object = read_file_object(cp, file).await?;
-    Ok(String::from_utf8_lossy(&read_object.object.bytes).to_string())
-}
+// moved read_file_content -> files.rs
 
-async fn read_file_output(cp: &ControlPlane, file: &resources_proto::File) -> Result<ToolOutput> {
-    let read_object = read_file_object(cp, file).await?;
-    let spec_media_type = file
-        .spec
-        .as_ref()
-        .map(|spec| spec.media_type.trim())
-        .unwrap_or_default();
-    let object_media_type = read_object.object.metadata.media_type.trim();
-    let media_type = if !spec_media_type.is_empty() {
-        spec_media_type.to_string()
-    } else if !object_media_type.is_empty() {
-        object_media_type.to_string()
-    } else {
-        file.spec
-            .as_ref()
-            .and_then(|spec| mime_guess::from_path(&spec.path).first_raw())
-            .unwrap_or("application/octet-stream")
-            .to_string()
-    };
-    let filename = {
-        let metadata_filename = read_object.object.metadata.filename.trim();
-        if !metadata_filename.is_empty() {
-            metadata_filename.to_string()
-        } else {
-            file.spec
-                .as_ref()
-                .map(|spec| {
-                    spec.path
-                        .trim_end_matches('/')
-                        .rsplit('/')
-                        .next()
-                        .unwrap_or_default()
-                        .to_string()
-                })
-                .unwrap_or_default()
-        }
-    };
-    let mut object_ref = read_object.object_ref;
-    object_ref.media_type = media_type.clone();
-    object_ref.filename = filename.clone();
-    Ok(ToolOutput::from_source_object(
-        read_object.object.bytes,
-        media_type,
-        filename,
-        object_ref,
-    ))
-}
+// moved read_file_output -> files.rs
 
-async fn read_file_object(
-    cp: &ControlPlane,
-    file: &resources_proto::File,
-) -> Result<ReadFileObject> {
-    let object_ref = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref())
-        .ok_or_else(|| anyhow!("File has no objectRef"))?;
-    let object = crate::control::cas::CasStore::new(cp.objects.clone())
-        .get_object_decoded(&object_ref.key)
-        .await?
-        .ok_or_else(|| anyhow!("File object '{}' not found", object_ref.key))?;
-    let object_ref = crate::control::cas::object_ref_from_stored_object(&object_ref.key, &object);
-    Ok(ReadFileObject { object, object_ref })
-}
+// moved read_file_object -> files.rs
 
-async fn upsert_memory_file(
-    cp: &ControlPlane,
-    namespace: &str,
-    path: &str,
-    media_type: &str,
-    content: &[u8],
-) -> Result<resources_proto::File> {
-    let existing = find_memory_file_by_path(cp, namespace, path).await?;
-    upsert_file(
-        cp,
-        namespace,
-        existing,
-        path,
-        media_type,
-        resources_proto::FilePurpose::Memory as i32,
-        resources_proto::FileIndexPolicy::Retrieval as i32,
-        resources_proto::FileRetention::Retained as i32,
-        content,
-    )
-    .await
-}
 
-async fn upsert_file(
-    cp: &ControlPlane,
-    namespace: &str,
-    existing: Option<resources_proto::File>,
-    path: &str,
-    media_type: &str,
-    purpose: i32,
-    index_policy: i32,
-    retention: i32,
-    content: &[u8],
-) -> Result<resources_proto::File> {
-    let store = ResourceStore::new(cp.kv.clone(), cp.pubsub.clone());
-    let name = existing
-        .as_ref()
-        .map(file_name_from_file)
-        .filter(|name| !name.is_empty())
-        .unwrap_or_else(|| keys::file_name_for_path(path));
-    let status = existing
-        .as_ref()
-        .and_then(|file| file.status.clone())
-        .unwrap_or_default();
-    let spec = resources_proto::FileSpec {
-        path: path.to_string(),
-        media_type: media_type.to_string(),
-        purpose,
-        index_policy,
-        retention,
-    };
-    let mut resource = store
-        .upsert(
-            namespace,
-            resource_model::file_resource(
-                namespace.to_string(),
-                name,
-                spec,
-                status,
-                file_resource_labels(purpose, index_policy, retention),
-            ),
-        )
-        .await?;
-    let uid = resource
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.uid.as_str())
-        .filter(|uid| !uid.is_empty())
-        .ok_or_else(|| anyhow!("File resource uid missing after upsert"))?;
-    let object_ref = write_file_objects(cp, namespace, uid, path, media_type, content).await?;
-    let status = resources_proto::FileStatus {
-        observed_generation: resource
-            .metadata
-            .as_ref()
-            .map(|metadata| metadata.generation)
-            .unwrap_or_default(),
-        phase: "Ready".to_string(),
-        conditions: Vec::new(),
-        object_ref: Some(object_ref),
-        updated_at: chrono::Utc::now().timestamp_micros(),
-        pending_upload: None,
-    };
-    resource.status = Some(resources_proto::ResourceStatus {
-        kind: Some(resources_proto::resource_status::Kind::File(status)),
-    });
-    let name = resource
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.name.clone())
-        .ok_or_else(|| anyhow!("File resource name missing"))?;
-    let resource = store
-        .patch_status(namespace, "File", &name, None, resource.status.unwrap())
-        .await?;
-    file_from_resource(resource).ok_or_else(|| anyhow!("invalid File resource"))
-}
+// moved upsert_file -> files.rs
 
-async fn write_file_objects(
-    cp: &ControlPlane,
-    namespace: &str,
-    file_uid: &str,
-    path: &str,
-    media_type: &str,
-    content: &[u8],
-) -> Result<resources_proto::FileObjectRef> {
-    let cas = crate::control::cas::CasStore::new(cp.objects.clone());
-    let object_ref = cas
-        .put_file(namespace, file_uid, path, content, media_type)
-        .await?;
-    cas.put_latest_file(namespace, path, content, media_type)
-        .await?;
-    Ok(resources_proto::FileObjectRef {
-        key: object_ref.key,
-        media_type: object_ref.media_type,
-        size_bytes: object_ref.size_bytes,
-        sha256: object_ref.sha256,
-        filename: object_ref.filename,
-        metadata: object_ref.metadata,
-    })
-}
+// moved write_file_objects -> files.rs
 
-fn file_from_resource(resource: resources_proto::Resource) -> Option<resources_proto::File> {
-    let spec = resource.spec.and_then(|spec| match spec.kind {
-        Some(resources_proto::resource_spec::Kind::File(spec)) => Some(spec),
-        _ => None,
-    })?;
-    let status = resource.status.and_then(|status| match status.kind {
-        Some(resources_proto::resource_status::Kind::File(status)) => Some(status),
-        _ => None,
-    });
-    Some(resources_proto::File {
-        metadata: resource.metadata,
-        spec: Some(spec),
-        status,
-    })
-}
+// moved file_from_resource -> files.rs
 
-fn file_name_from_file(file: &resources_proto::File) -> String {
-    file.metadata
-        .as_ref()
-        .map(|metadata| metadata.name.clone())
-        .unwrap_or_default()
-}
+// moved file_name_from_file -> files.rs
 
-fn file_json(file: &resources_proto::File, include_resource_name: bool) -> Value {
-    let namespace = file
-        .metadata
-        .as_ref()
-        .map(|metadata| metadata.namespace.as_str())
-        .unwrap_or_default();
-    let spec = file.spec.as_ref();
-    let path = spec.map(|spec| spec.path.as_str()).unwrap_or_default();
-    let object = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref());
-    let mut value = json!({
-        "namespace": namespace,
-        "uri": file_uri(namespace, path),
-        "path": path,
-        "mediaType": spec.map(|spec| spec.media_type.as_str()).unwrap_or_default(),
-        "purpose": spec.map(|spec| file_purpose_label(spec.purpose)).unwrap_or("unspecified"),
-        "indexPolicy": spec.map(|spec| file_index_policy_label(spec.index_policy)).unwrap_or("unspecified"),
-        "retention": spec.map(|spec| file_retention_label(spec.retention)).unwrap_or("unspecified"),
-        "sizeBytes": object.map(|object| object.size_bytes).unwrap_or_default(),
-        "sha256": object.map(|object| object.sha256.as_str()).unwrap_or_default(),
-    });
-    if include_resource_name {
-        if let Some(map) = value.as_object_mut() {
-            map.insert("name".to_string(), json!(file_name_from_file(file)));
-        }
-    }
-    value
-}
+// moved file_json -> files.rs
 
-fn memory_file_json(file: &resources_proto::File) -> Value {
-    let spec = file.spec.as_ref();
-    let object = file
-        .status
-        .as_ref()
-        .and_then(|status| status.object_ref.as_ref());
-    json!({
-        "namespace": file.metadata.as_ref().map(|metadata| metadata.namespace.as_str()).unwrap_or_default(),
-        "name": file_name_from_file(file),
-        "path": spec.map(|spec| spec.path.as_str()).unwrap_or_default(),
-        "mediaType": spec.map(|spec| spec.media_type.as_str()).unwrap_or_default(),
-        "purpose": "MEMORY",
-        "indexPolicy": "RETRIEVAL",
-        "sizeBytes": object.map(|object| object.size_bytes).unwrap_or_default(),
-        "sha256": object.map(|object| object.sha256.as_str()).unwrap_or_default(),
-    })
-}
 
-fn file_uri(namespace: &str, path: &str) -> String {
-    format!("file://{}{}", namespace, path)
-}
+// moved file_uri -> files.rs
 
-fn file_location_from_args(current_namespace: &str, args: &Value) -> Result<(String, String)> {
-    if let Some(uri) = opt_str(args, "uri") {
-        return parse_file_uri(uri);
-    }
-    let namespace = namespace_arg(current_namespace, args);
-    let path = normalize_logical_path(req_str(args, "path")?)?;
-    Ok((namespace, path))
-}
+// moved file_location_from_args -> files.rs
 
-fn namespace_arg(current_namespace: &str, args: &Value) -> String {
-    match opt_str(args, "namespace") {
-        Some("current" | "current_namespace" | "current namespace") | None => {
-            current_namespace.to_string()
-        }
-        Some(namespace) => namespace.to_string(),
-    }
-}
+// moved namespace_arg -> files.rs
 
-fn parse_file_uri(uri: &str) -> Result<(String, String)> {
-    let rest = uri
-        .trim()
-        .strip_prefix("file://")
-        .ok_or_else(|| anyhow!("file uri must start with 'file://'"))?;
-    let split = rest
-        .find('/')
-        .ok_or_else(|| anyhow!("file uri must include namespace and path"))?;
-    let namespace = rest[..split].trim();
-    if namespace.is_empty() {
-        return Err(anyhow!("file uri namespace is required"));
-    }
-    let path = normalize_logical_path(&rest[split..])?;
-    Ok((namespace.to_string(), path))
-}
+// moved parse_file_uri -> files.rs
 
-pub(super) fn ensure_file_read_namespace(
-    current_namespace: &str,
-    requested_namespace: &str,
-) -> Result<()> {
-    let requested_namespace = if requested_namespace == "current" {
-        current_namespace
-    } else {
-        requested_namespace
-    };
-    if crate::control::ns::ancestry(current_namespace)
-        .iter()
-        .any(|namespace| namespace == requested_namespace)
-    {
-        Ok(())
-    } else {
-        Err(anyhow!(
-            "file access to namespace '{}' is outside the current namespace ancestry",
-            requested_namespace
-        ))
-    }
-}
+// moved ensure_file_read_namespace -> files.rs
 
-fn normalize_memory_path(path: &str) -> Result<String> {
-    let path = normalize_logical_path(path)?;
-    if !path.starts_with("/memory/") && path != "/memory" {
-        return Err(anyhow!("memory path must be under /memory"));
-    }
-    Ok(path)
-}
 
-fn file_resource_labels(
-    purpose: i32,
-    index_policy: i32,
-    retention: i32,
-) -> HashMap<String, String> {
-    HashMap::from([
-        (
-            "talon.impalasys.com/purpose".to_string(),
-            file_purpose_label(purpose).to_string(),
-        ),
-        (
-            "talon.impalasys.com/index-policy".to_string(),
-            file_index_policy_label(index_policy).to_string(),
-        ),
-        (
-            "talon.impalasys.com/retention".to_string(),
-            file_retention_label(retention).to_string(),
-        ),
-    ])
-}
+// moved file_resource_labels -> files.rs
 
-fn file_purpose_label(value: i32) -> &'static str {
-    match resources_proto::FilePurpose::try_from(value).ok() {
-        Some(resources_proto::FilePurpose::Memory) => "memory",
-        Some(resources_proto::FilePurpose::Artifact) => "artifact",
-        Some(resources_proto::FilePurpose::Skill) => "skill",
-        _ => "unspecified",
-    }
-}
+// moved file_purpose_label -> files.rs
 
-fn file_index_policy_label(value: i32) -> &'static str {
-    match resources_proto::FileIndexPolicy::try_from(value).ok() {
-        Some(resources_proto::FileIndexPolicy::None) => "none",
-        Some(resources_proto::FileIndexPolicy::Search) => "search",
-        Some(resources_proto::FileIndexPolicy::Retrieval) => "retrieval",
-        _ => "unspecified",
-    }
-}
+// moved file_index_policy_label -> files.rs
 
-fn file_retention_label(value: i32) -> &'static str {
-    match resources_proto::FileRetention::try_from(value).ok() {
-        Some(resources_proto::FileRetention::Retained) => "retained",
-        _ => "unspecified",
-    }
-}
+// moved file_retention_label -> files.rs
 
-fn parse_file_purpose(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "artifact" | "file-purpose-artifact" => Ok(resources_proto::FilePurpose::Artifact as i32),
-        "memory" | "file-purpose-memory" => Ok(resources_proto::FilePurpose::Memory as i32),
-        "skill" | "file-purpose-skill" => Ok(resources_proto::FilePurpose::Skill as i32),
-        _ => Err(anyhow!(
-            "unsupported File purpose '{}'; expected ARTIFACT, MEMORY, or SKILL",
-            value
-        )),
-    }
-}
+// moved parse_file_purpose -> files.rs
 
-fn parse_file_index_policy(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "none" | "file-index-policy-none" => Ok(resources_proto::FileIndexPolicy::None as i32),
-        "search" | "file-index-policy-search" => {
-            Ok(resources_proto::FileIndexPolicy::Search as i32)
-        }
-        "retrieval" | "file-index-policy-retrieval" => {
-            Ok(resources_proto::FileIndexPolicy::Retrieval as i32)
-        }
-        _ => Err(anyhow!(
-            "unsupported File index_policy '{}'; expected NONE, SEARCH, or RETRIEVAL",
-            value
-        )),
-    }
-}
+// moved parse_file_index_policy -> files.rs
 
-fn parse_file_retention(value: &str) -> Result<i32> {
-    let normalized = normalize_enum_input(value);
-    match normalized.as_str() {
-        "retained" | "file-retention-retained" => {
-            Ok(resources_proto::FileRetention::Retained as i32)
-        }
-        _ => Err(anyhow!(
-            "unsupported File retention '{}'; expected RETAINED",
-            value
-        )),
-    }
-}
+// moved parse_file_retention -> files.rs
 
-fn normalize_enum_input(value: &str) -> String {
-    value.trim().to_ascii_lowercase().replace(['_', ' '], "-")
-}
+// moved normalize_enum_input -> files.rs
 
-fn memory_excerpt(content: &str, query: &str) -> String {
-    let lower = content.to_ascii_lowercase();
-    let query_index = lower.find(query).unwrap_or(0);
-    let mut byte_count: usize = 0;
-    let mut char_index: usize = 0;
-    for ch in content.chars() {
-        if byte_count >= query_index {
-            break;
-        }
-        byte_count += ch.len_utf8();
-        char_index += 1;
-    }
-    let start = char_index.saturating_sub(120);
-    content
-        .chars()
-        .skip(start)
-        .take(340)
-        .collect::<String>()
-        .trim()
-        .to_string()
-}
 
-async fn create_artifact(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<String> {
-    if current_session.trim().is_empty() {
-        return Err(anyhow!("create_artifact requires an active session"));
-    }
-    let title = req_str(args, "title")?;
-    let media_type = opt_str(args, "media_type").unwrap_or("text/markdown");
-    if args.get("content").and_then(Value::as_str).is_none()
-        && opt_str(args, "content_base64").is_none()
-    {
-        return Err(anyhow!(
-            "create_artifact requires content or content_base64"
-        ));
-    }
-    let content = artifact_content_bytes(args)?;
-    let labels = string_map(args.get("labels"));
-    let metadata = string_map(args.get("metadata"));
-    let artifact_id = crate::control::uuid::unique_name("artifact");
-    let object_ref = crate::control::cas::CasStore::new(cp.objects.clone())
-        .put_artifact(
-            current_namespace,
-            current_agent,
-            current_session,
-            &artifact_id,
-            &content,
-            media_type,
-            metadata.clone(),
-        )
-        .await?;
-    let artifact = crate::gateway::rpc::data_proto::Artifact {
-        id: artifact_id.clone(),
-        session_id: current_session.to_string(),
-        title: title.to_string(),
-        media_type: media_type.to_string(),
-        object_ref: Some(object_ref),
-        created_by_agent: current_agent.to_string(),
-        created_at: chrono::Utc::now().timestamp_micros(),
-        labels,
-        metadata,
-    };
-    cp.kv
-        .set_msg(
-            &keys::artifact(
-                current_namespace,
-                current_agent,
-                current_session,
-                &artifact_id,
-            ),
-            &artifact,
-        )
-        .await?;
-    let artifact_uri = ArtifactUri {
-        namespace: current_namespace.to_string(),
-        agent: current_agent.to_string(),
-        session_id: current_session.to_string(),
-        artifact_id: artifact_id.clone(),
-    }
-    .encode();
-    Ok(serde_json::to_string_pretty(&json!({
-        "artifact": artifact_json(&artifact),
-        "artifactUri": artifact_uri
-    }))?)
-}
+// moved create_artifact -> artifacts.rs
 
-async fn read_artifact(
-    cp: &ControlPlane,
-    _current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<ToolOutput> {
-    let artifact_uri = req_str(args, "artifact_uri")?;
-    let (_, artifact) =
-        resolve_artifact_uri(cp, current_agent, current_session, artifact_uri, OP_READ).await?;
-    let object_ref = artifact
-        .object_ref
-        .as_ref()
-        .ok_or_else(|| anyhow!("Artifact has no objectRef"))?;
-    let object = cp
-        .objects
-        .get(&object_ref.key)
-        .await?
-        .ok_or_else(|| anyhow!("Artifact object not found"))?;
-    let media_type = if !artifact.media_type.trim().is_empty() {
-        artifact.media_type.trim().to_string()
-    } else if !object_ref.media_type.trim().is_empty() {
-        object_ref.media_type.trim().to_string()
-    } else if !object.metadata.media_type.trim().is_empty() {
-        object.metadata.media_type.trim().to_string()
-    } else {
-        "application/octet-stream".to_string()
-    };
-    let filename = if !object_ref.filename.trim().is_empty() {
-        object_ref.filename.clone()
-    } else if !object.metadata.filename.trim().is_empty() {
-        object.metadata.filename.clone()
-    } else {
-        artifact.title.clone()
-    };
-    let mut object_ref = object_ref.clone();
-    object_ref.media_type = media_type.clone();
-    object_ref.filename = filename.clone();
-    Ok(ToolOutput::from_source_object(
-        object.bytes,
-        media_type,
-        filename,
-        object_ref,
-    ))
-}
+// moved read_artifact -> artifacts.rs
 
-async fn update_artifact(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<String> {
-    let artifact_uri = req_str(args, "artifact_uri")?;
-    let uri = parse_artifact_uri(artifact_uri)?;
-    if current_namespace != uri.namespace
-        || current_agent != uri.agent
-        || current_session != uri.session_id
-    {
-        return Err(anyhow!(
-            "only the owning artifact namespace/agent/session may update '{artifact_uri}'"
-        ));
-    }
-    let mut artifact = cp
-        .kv
-        .get_msg::<crate::gateway::rpc::data_proto::Artifact>(&keys::artifact(
-            &uri.namespace,
-            &uri.agent,
-            &uri.session_id,
-            &uri.artifact_id,
-        ))
-        .await?
-        .ok_or_else(|| anyhow!("Artifact '{}' not found", uri.artifact_id))?;
-    let previous_object_key = artifact
-        .object_ref
-        .as_ref()
-        .map(|object_ref| object_ref.key.clone());
-    let media_type = opt_str(args, "media_type").unwrap_or(&artifact.media_type);
-    if args.get("content").and_then(Value::as_str).is_none()
-        && opt_str(args, "content_base64").is_none()
-    {
-        return Err(anyhow!(
-            "update_artifact requires content or content_base64"
-        ));
-    }
-    let content = artifact_content_bytes(args)?;
-    let cas = crate::control::cas::CasStore::new(cp.objects.clone());
-    let object_ref = cas
-        .put_artifact(
-            &uri.namespace,
-            &uri.agent,
-            &uri.session_id,
-            &uri.artifact_id,
-            &content,
-            media_type,
-            artifact.metadata.clone(),
-        )
-        .await?;
-    artifact.media_type = media_type.to_string();
-    artifact.object_ref = Some(object_ref);
-    let artifact_key = keys::artifact(
-        &uri.namespace,
-        &uri.agent,
-        &uri.session_id,
-        &uri.artifact_id,
-    );
-    if let Err(error) = cp.kv.set_msg(&artifact_key, &artifact).await {
-        if let Some(new_object_key) = artifact
-            .object_ref
-            .as_ref()
-            .map(|object_ref| &object_ref.key)
-        {
-            if let Err(cleanup_error) = cas.delete_object(new_object_key).await {
-                tracing::warn!(
-                    error = %cleanup_error,
-                    object_key = %new_object_key,
-                    "failed to delete uncommitted artifact CAS object after update failure"
-                );
-            }
-        }
-        return Err(error);
-    }
-    if let Some(previous_object_key) = previous_object_key {
-        if artifact
-            .object_ref
-            .as_ref()
-            .is_none_or(|object_ref| object_ref.key != previous_object_key)
-        {
-            if let Err(error) = cas.delete_object(&previous_object_key).await {
-                tracing::warn!(
-                    error = %error,
-                    object_key = %previous_object_key,
-                    "failed to delete superseded artifact CAS object"
-                );
-            }
-        }
-    }
-    Ok(serde_json::to_string_pretty(&json!({
-        "artifact": artifact_json(&artifact),
-        "artifactUri": uri.encode()
-    }))?)
-}
+// moved update_artifact -> artifacts.rs
 
-async fn get_artifact_metadata(
-    cp: &ControlPlane,
-    _current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<String> {
-    let artifact_uri = req_str(args, "artifact_uri")?;
-    let (_, artifact) = resolve_artifact_uri(
-        cp,
-        current_agent,
-        current_session,
-        artifact_uri,
-        OP_METADATA,
-    )
-    .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "artifact": artifact_json(&artifact)
-    }))?)
-}
+// moved get_artifact_metadata -> artifacts.rs
 
-async fn grant_artifact(
-    cp: &ControlPlane,
-    _current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<String> {
-    let artifact_uri = req_str(args, "artifact_uri")?;
-    let (uri, _) =
-        resolve_artifact_uri(cp, current_agent, current_session, artifact_uri, OP_READ).await?;
-    let operations = args
-        .get("operations")
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(Value::as_str)
-                .map(str::to_string)
-                .collect::<Vec<_>>()
-        })
-        .filter(|values| !values.is_empty())
-        .unwrap_or_else(|| vec![OP_READ.to_string(), OP_METADATA.to_string()]);
-    for operation in &operations {
-        if !matches!(operation.as_str(), OP_READ | OP_METADATA | OP_PROMOTE) {
-            return Err(anyhow!("unsupported artifact operation '{}'", operation));
-        }
-    }
-    let ttl = args
-        .get("ttl_seconds")
-        .and_then(Value::as_i64)
-        .map(access_expiry_from_ttl_seconds)
-        .unwrap_or_else(default_access_expiry);
-    let target_agent = opt_str(args, "target_agent").unwrap_or("");
-    let target_session_id = opt_str(args, "target_session_id").unwrap_or("");
-    let access = crate::gateway::rpc::data_proto::ArtifactAccess {
-        target_agent: target_agent.to_string(),
-        target_session_id: target_session_id.to_string(),
-        operations,
-        expires_at: ttl,
-        granted_by_agent: current_agent.to_string(),
-        granted_by_session_id: current_session.to_string(),
-        created_at: chrono::Utc::now().timestamp_micros(),
-    };
-    cp.kv
-        .set_msg(
-            &keys::artifact_access(
-                &uri.namespace,
-                &uri.agent,
-                &uri.session_id,
-                &uri.artifact_id,
-                target_agent,
-                target_session_id,
-            ),
-            &access,
-        )
-        .await?;
-    Ok(serde_json::to_string_pretty(&json!({
-        "artifactUri": uri.encode()
-    }))?)
-}
+// moved grant_artifact -> artifacts.rs
 
-async fn fetch_url(args: &Value) -> Result<String> {
-    let url = req_str(args, "url")?;
-    let mut current_url = validate_public_http_url(url).await?;
-    let max_chars = args
-        .get("max_chars")
-        .and_then(Value::as_u64)
-        .unwrap_or(12_000)
-        .clamp(1_000, 40_000) as usize;
-    let mut redirects = 0usize;
-    let response = loop {
-        let client = http_client(&current_url)?;
-        let response = client.get(current_url.url.clone()).send().await?;
-        if !response.status().is_redirection() {
-            break response;
-        }
-        if redirects >= 5 {
-            return Err(anyhow!("too many redirects while fetching URL"));
-        }
-        let Some(location) = response.headers().get(reqwest::header::LOCATION) else {
-            break response;
-        };
-        let location = location
-            .to_str()
-            .map_err(|err| anyhow!("redirect Location is not valid UTF-8: {err}"))?;
-        current_url = validate_public_http_url(current_url.url.join(location)?.as_str()).await?;
-        redirects += 1;
-    };
-    let status = response.status().as_u16();
-    let final_url = response.url().to_string();
-    let content_type = response
-        .headers()
-        .get(reqwest::header::CONTENT_TYPE)
-        .and_then(|value| value.to_str().ok())
-        .unwrap_or("")
-        .to_string();
-    let body = response.text().await?;
-    let title = extract_title(&body);
-    let visible_text = compact_visible_text(&body, max_chars);
+// moved fetch_url -> research.rs
 
-    Ok(serde_json::to_string_pretty(&json!({
-        "url": url,
-        "finalUrl": final_url,
-        "status": status,
-        "contentType": content_type,
-        "title": title,
-        "text": visible_text,
-        "truncated": body.len() > max_chars
-    }))?)
-}
+// moved web_search -> research.rs
 
-async fn web_search(args: &Value) -> Result<String> {
-    let query = req_str(args, "query")?;
-    let limit = args
-        .get("limit")
-        .and_then(Value::as_u64)
-        .unwrap_or(5)
-        .clamp(1, 10) as usize;
-    let url = format!(
-        "https://duckduckgo.com/html/?q={}",
-        urlencoding::encode(query)
-    );
-    let url = validate_public_http_url(&url).await?;
-    let response = http_client(&url)?.get(url.url.clone()).send().await?;
-    let status = response.status().as_u16();
-    let body = response.text().await?;
-    let results = extract_duckduckgo_results(&body, limit);
+// moved upsert_schedule -> schedules.rs
 
-    Ok(serde_json::to_string_pretty(&json!({
-        "query": query,
-        "provider": "duckduckgo-html",
-        "status": status,
-        "results": results,
-        "citationPolicy": "Fetch a result URL with fetch_url before using it as evidence."
-    }))?)
-}
-
-async fn upsert_schedule(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    args: &Value,
-    existing: Option<resources_proto::Schedule>,
-) -> Result<resources_proto::Schedule> {
-    let namespace = opt_str(args, "namespace")
-        .unwrap_or(current_namespace)
-        .to_string();
-    let name = req_str(args, "name")?.to_string();
-    let existing_spec = existing
-        .as_ref()
-        .and_then(|schedule| schedule.spec.as_ref());
-    let existing_target = existing_spec.and_then(|spec| spec.target.as_ref());
-    let kind = scheduling::normalize_schedule_kind(
-        opt_str(args, "kind")
-            .or_else(|| existing_spec.map(|spec| spec.kind.as_str()))
-            .unwrap_or(""),
-    );
-    let cron = opt_str(args, "cron")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.cron.clone()))
-        .unwrap_or_default();
-    let interval_seconds = args
-        .get("interval_seconds")
-        .and_then(Value::as_u64)
-        .map(|value| value as u32)
-        .or_else(|| existing_spec.map(|spec| spec.interval_seconds))
-        .unwrap_or_default();
-    let run_at = opt_str(args, "run_at")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.run_at.clone()))
-        .unwrap_or_default();
-    let timezone = opt_str(args, "timezone")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.timezone.clone()))
-        .unwrap_or_default();
-    let agent = opt_str(args, "agent")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.agent.clone()))
-        .unwrap_or_else(|| current_agent.to_string());
-    let session_mode = opt_str(args, "session_mode")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.session_mode.clone()))
-        .unwrap_or_else(|| "new".to_string());
-    let session_mode = scheduling::normalize_session_mode(&session_mode)?;
-    let session_id = opt_str(args, "session_id")
-        .map(str::to_string)
-        .or_else(|| existing_target.map(|target| target.session_id.clone()))
-        .unwrap_or_default();
-    let input_message = opt_str(args, "input_message")
-        .map(str::to_string)
-        .or_else(|| existing_spec.map(|spec| spec.input_message.clone()))
-        .unwrap_or_default();
-    let enabled = args
-        .get("enabled")
-        .and_then(Value::as_bool)
-        .or_else(|| existing_spec.map(|spec| spec.enabled))
-        .unwrap_or(true);
-    let labels = args
-        .get("labels")
-        .and_then(Value::as_object)
-        .map(|map| {
-            map.iter()
-                .filter_map(|(key, value)| {
-                    value
-                        .as_str()
-                        .map(|current| (key.clone(), current.to_string()))
-                })
-                .collect::<HashMap<_, _>>()
-        })
-        .or_else(|| existing.as_ref().map(|schedule| schedule.labels().clone()))
-        .unwrap_or_default();
-
-    let mut schedule = resource_model::schedule(
-        namespace.clone(),
-        name.clone(),
-        resources_proto::ScheduleSpec {
-            kind,
-            cron,
-            interval_seconds,
-            run_at,
-            timezone,
-            target: Some(resources_proto::ScheduleTarget {
-                agent,
-                workflow: String::new(),
-                session_mode,
-                session_id,
-            }),
-            input_message,
-            input_json: String::new(),
-            enabled,
-        },
-        existing
-            .and_then(|schedule| schedule.status)
-            .unwrap_or_default(),
-        labels,
-    );
-
-    scheduling::initialize_schedule(&mut schedule, chrono::Utc::now())?;
-    let next_run = schedule
-        .status
-        .as_ref()
-        .and_then(|status| status.next_run_at)
-        .and_then(chrono::DateTime::from_timestamp_micros);
-    scheduling::persist_schedule(cp.kv.as_ref(), &schedule).await?;
-    scheduling::arm_schedule(cp.scheduler.as_ref(), &mut schedule, next_run).await?;
-    scheduling::persist_schedule(cp.kv.as_ref(), &schedule).await?;
-    Ok(schedule)
-}
-
-fn schedule_json(schedule: &resources_proto::Schedule) -> Value {
-    let spec = schedule.spec.as_ref();
-    let status = schedule.status.as_ref();
-    let target = spec.and_then(|spec| spec.target.as_ref());
-    json!({
-        "name": schedule.name(),
-        "ns": schedule.namespace(),
-        "spec": {
-            "kind": spec.map(|spec| spec.kind.clone()).unwrap_or_default(),
-            "cron": spec.map(|spec| spec.cron.clone()).unwrap_or_default(),
-            "intervalSeconds": spec.map(|spec| spec.interval_seconds).unwrap_or_default(),
-            "runAt": spec.map(|spec| spec.run_at.clone()).unwrap_or_default(),
-            "timezone": spec.map(|spec| spec.timezone.clone()).unwrap_or_default(),
-            "target": {
-                "agent": target.map(|target| target.agent.clone()).unwrap_or_default(),
-                "sessionMode": target.map(|target| target.session_mode.clone()).unwrap_or_default(),
-                "sessionId": target.map(|target| target.session_id.clone()).unwrap_or_default(),
-            },
-            "inputMessage": spec.map(|spec| spec.input_message.clone()).unwrap_or_default(),
-            "enabled": spec.map(|spec| spec.enabled).unwrap_or(false),
-        },
-        "status": status.map(|status| json!({
-            "revision": status.revision,
-            "backendArmed": status.backend_armed,
-            "backendHandle": status.backend_handle,
-            "nextRunAt": status.next_run_at,
-            "lastRunAt": status.last_run_at,
-            "lastSessionId": status.last_session_id,
-            "lastError": status.last_error,
-            "claimedRunAt": status.claimed_run_at,
-            "claimExpiresAt": status.claim_expires_at,
-            "recentEvents": status.recent_events.iter().map(|event| json!({
-                "timestamp": event.timestamp,
-                "phase": event.phase,
-                "outcome": event.outcome,
-                "detail": event.detail,
-            })).collect::<Vec<_>>()
-        })).unwrap_or_else(|| json!({})),
-        "labels": schedule.labels(),
-    })
-}
+// moved schedule_json -> schedules.rs
 
 fn create_task(
     current_namespace: &str,
@@ -3073,334 +809,43 @@ fn is_terminal_phase(value: i32) -> bool {
     )
 }
 
-async fn create_goal(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<data_proto::Goal> {
-    let objective = req_str(args, "objective")?.to_string();
-    let now = chrono::Utc::now().timestamp_micros();
-    let goal = data_proto::Goal {
-        id: crate::control::uuid::unique_name("goal"),
-        namespace: current_namespace.to_string(),
-        agent: current_agent.to_string(),
-        session_id: current_session.to_string(),
-        objective,
-        success_criteria: string_vec(args.get("success_criteria")),
-        phase: data_proto::GoalPhase::Running as i32,
-        progress_summary: opt_str(args, "progress_summary")
-            .unwrap_or("Goal created.")
-            .to_string(),
-        iteration: 0,
-        max_iterations: args
-            .get("max_iterations")
-            .and_then(Value::as_i64)
-            .unwrap_or_default()
-            .try_into()
-            .unwrap_or_default(),
-        created_at: now,
-        updated_at: now,
-        completed_at: 0,
-        blocked_reason: String::new(),
-        labels: string_map(args.get("labels")),
-        metadata: string_map(args.get("metadata")),
-    };
-    upsert_goal(cp, goal.clone()).await?;
-    Ok(goal)
-}
+// moved create_goal -> goals.rs
 
-async fn get_goal_from_args(
-    cp: &ControlPlane,
-    current_namespace: &str,
-    current_agent: &str,
-    current_session: &str,
-    args: &Value,
-) -> Result<data_proto::Goal> {
-    let namespace = opt_str(args, "namespace").unwrap_or(current_namespace);
-    let agent = opt_str(args, "agent").unwrap_or(current_agent);
-    let session_id = opt_str(args, "session_id").unwrap_or(current_session);
-    let goal_id = req_str(args, "goal_id")?;
-    load_goal(cp, namespace, agent, session_id, goal_id)
-        .await?
-        .ok_or_else(|| anyhow!("goal '{}' not found", goal_id))
-}
+// moved get_goal_from_args -> goals.rs
 
-async fn load_goal(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    goal_id: &str,
-) -> Result<Option<data_proto::Goal>> {
-    cp.kv
-        .get_msg::<data_proto::Goal>(&keys::goal(namespace, agent, session_id, goal_id))
-        .await
-}
+// moved load_goal -> goals.rs
 
-async fn upsert_goal(cp: &ControlPlane, goal: data_proto::Goal) -> Result<()> {
-    cp.kv
-        .set_msg(
-            &keys::goal(&goal.namespace, &goal.agent, &goal.session_id, &goal.id),
-            &goal,
-        )
-        .await
-}
+// moved upsert_goal -> goals.rs
 
-async fn list_goals(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    status_group: Option<&str>,
-    phase: Option<&str>,
-    limit: usize,
-) -> Result<Vec<data_proto::Goal>> {
-    list_session_goals(cp, namespace, agent, session_id, status_group, phase, limit).await
-}
+// moved list_goals -> goals.rs
 
-async fn list_session_goals(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-    status_group: Option<&str>,
-    phase: Option<&str>,
-    limit: usize,
-) -> Result<Vec<data_proto::Goal>> {
-    let mut goals = cp
-        .kv
-        .list_entries(&keys::goal_prefix(namespace, agent, session_id), None)
-        .await?
-        .into_iter()
-        .filter_map(|(_, value)| data_proto::Goal::decode(value.as_slice()).ok())
-        .filter(|goal| goal_matches(goal, status_group, phase))
-        .collect::<Vec<_>>();
-    goals.sort_by(|left, right| right.updated_at.cmp(&left.updated_at));
-    goals.truncate(limit);
-    Ok(goals)
-}
+// moved list_session_goals -> goals.rs
 
-fn goal_matches(goal: &data_proto::Goal, status_group: Option<&str>, phase: Option<&str>) -> bool {
-    if status_group
-        .is_some_and(|current| !goal_status_group(goal.phase).eq_ignore_ascii_case(current))
-    {
-        return false;
-    }
-    if phase.is_some_and(|current| parse_goal_phase(current).ok() != Some(goal.phase)) {
-        return false;
-    }
-    true
-}
+// moved goal_matches -> goals.rs
 
-fn update_goal_from_args(goal: &mut data_proto::Goal, args: &Value) -> Result<()> {
-    let now = chrono::Utc::now().timestamp_micros();
-    if let Some(phase) = opt_str(args, "phase") {
-        goal.phase = parse_goal_phase(phase)?;
-    }
-    if let Some(summary) = opt_str(args, "progress_summary") {
-        goal.progress_summary = summary.to_string();
-    }
-    if let Some(iteration) = args.get("iteration").and_then(Value::as_i64) {
-        goal.iteration = iteration.try_into().unwrap_or_default();
-    }
-    if let Some(blocked_reason) = opt_str(args, "blocked_reason") {
-        goal.blocked_reason = blocked_reason.to_string();
-    }
-    goal.updated_at = now;
-    if is_terminal_goal_phase(goal.phase) && goal.completed_at == 0 {
-        goal.completed_at = now;
-    }
-    Ok(())
-}
+// moved update_goal_from_args -> goals.rs
 
-fn goal_json(goal: &data_proto::Goal) -> Value {
-    json!({
-        "id": goal.id,
-        "namespace": goal.namespace,
-        "agent": goal.agent,
-        "sessionId": goal.session_id,
-        "objective": goal.objective,
-        "successCriteria": goal.success_criteria,
-        "phase": goal_phase_name(goal.phase),
-        "statusGroup": goal_status_group(goal.phase),
-        "progressSummary": goal.progress_summary,
-        "iteration": goal.iteration,
-        "maxIterations": goal.max_iterations,
-        "createdAt": goal.created_at,
-        "updatedAt": goal.updated_at,
-        "completedAt": goal.completed_at,
-        "blockedReason": goal.blocked_reason,
-        "labels": goal.labels,
-        "metadata": goal.metadata,
-    })
-}
+// moved goal_json -> goals.rs
 
-pub async fn active_goals_context(
-    cp: &ControlPlane,
-    namespace: &str,
-    agent: &str,
-    session_id: &str,
-) -> Result<Option<String>> {
-    let goals = list_goals(cp, namespace, agent, session_id, Some("active"), None, 20).await?;
-    if goals.is_empty() {
-        return Ok(None);
-    }
+// moved active_goals_context -> goals.rs
 
-    let mut lines = vec![
-        "# Active Talon Goals".to_string(),
-        "Keep these session-scoped objectives in view while deciding next steps.".to_string(),
-    ];
-    for goal in goals {
-        lines.push(format!(
-            "- {} [{}] {}",
-            goal.id,
-            goal_phase_name(goal.phase),
-            goal.objective
-        ));
-        if !goal.success_criteria.is_empty() {
-            lines.push(format!(
-                "  Success criteria: {}",
-                goal.success_criteria.join("; ")
-            ));
-        }
-        if !goal.progress_summary.is_empty() {
-            lines.push(format!("  Progress: {}", goal.progress_summary));
-        }
-        if !goal.blocked_reason.is_empty() {
-            lines.push(format!("  Blocked reason: {}", goal.blocked_reason));
-        }
-    }
-    Ok(Some(lines.join("\n")))
-}
+// moved parse_goal_phase -> goals.rs
 
-fn parse_goal_phase(value: &str) -> Result<i32> {
-    let phase = match value.trim().to_ascii_uppercase().as_str() {
-        "" | "UNSPECIFIED" => data_proto::GoalPhase::Unspecified,
-        "RUNNING" => data_proto::GoalPhase::Running,
-        "PAUSED" => data_proto::GoalPhase::Paused,
-        "NEEDS_REVIEW" | "NEEDS-REVIEW" => data_proto::GoalPhase::NeedsReview,
-        "SUCCEEDED" | "SUCCESS" | "COMPLETED" => data_proto::GoalPhase::Succeeded,
-        "FAILED" => data_proto::GoalPhase::Failed,
-        "BLOCKED" => data_proto::GoalPhase::Blocked,
-        "CANCELED" | "CANCELLED" => data_proto::GoalPhase::Canceled,
-        "EXPIRED" => data_proto::GoalPhase::Expired,
-        other => return Err(anyhow!("unsupported goal phase '{}'", other)),
-    };
-    Ok(phase as i32)
-}
+// moved goal_phase_name -> goals.rs
 
-fn goal_phase_name(value: i32) -> &'static str {
-    match data_proto::GoalPhase::try_from(value).ok() {
-        Some(data_proto::GoalPhase::Running) => "RUNNING",
-        Some(data_proto::GoalPhase::Paused) => "PAUSED",
-        Some(data_proto::GoalPhase::NeedsReview) => "NEEDS_REVIEW",
-        Some(data_proto::GoalPhase::Succeeded) => "SUCCEEDED",
-        Some(data_proto::GoalPhase::Failed) => "FAILED",
-        Some(data_proto::GoalPhase::Blocked) => "BLOCKED",
-        Some(data_proto::GoalPhase::Canceled) => "CANCELED",
-        Some(data_proto::GoalPhase::Expired) => "EXPIRED",
-        _ => "UNSPECIFIED",
-    }
-}
+// moved goal_status_group -> goals.rs
 
-fn goal_status_group(value: i32) -> &'static str {
-    if is_active_goal_phase(value) {
-        "ACTIVE"
-    } else if is_terminal_goal_phase(value) {
-        "TERMINAL"
-    } else {
-        "UNKNOWN"
-    }
-}
+// moved is_active_goal_phase -> goals.rs
 
-fn is_active_goal_phase(value: i32) -> bool {
-    matches!(
-        data_proto::GoalPhase::try_from(value).ok(),
-        Some(data_proto::GoalPhase::Running)
-            | Some(data_proto::GoalPhase::Paused)
-            | Some(data_proto::GoalPhase::NeedsReview)
-            | Some(data_proto::GoalPhase::Blocked)
-    )
-}
+// moved is_terminal_goal_phase -> goals.rs
 
-fn is_terminal_goal_phase(value: i32) -> bool {
-    matches!(
-        data_proto::GoalPhase::try_from(value).ok(),
-        Some(data_proto::GoalPhase::Succeeded)
-            | Some(data_proto::GoalPhase::Failed)
-            | Some(data_proto::GoalPhase::Canceled)
-            | Some(data_proto::GoalPhase::Expired)
-    )
-}
 
-fn req_str<'a>(args: &'a Value, key: &str) -> Result<&'a str> {
-    args.get(key)
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty())
-        .ok_or_else(|| anyhow!("'{}' is required", key))
-}
 
-fn opt_str<'a>(args: &'a Value, key: &str) -> Option<&'a str> {
-    args.get(key)
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-}
 
-fn opt_usize(args: &Value, key: &str) -> Option<usize> {
-    args.get(key)
-        .and_then(Value::as_u64)
-        .map(|value| value as usize)
-}
 
-fn opt_u64(args: &Value, key: &str) -> Option<u64> {
-    args.get(key).and_then(Value::as_u64)
-}
 
-fn string_map(value: Option<&Value>) -> HashMap<String, String> {
-    value
-        .and_then(Value::as_object)
-        .map(|map| {
-            map.iter()
-                .filter_map(|(key, value)| {
-                    value
-                        .as_str()
-                        .map(|current| (key.clone(), current.to_string()))
-                })
-                .collect::<HashMap<_, _>>()
-        })
-        .unwrap_or_default()
-}
 
-fn string_vec(value: Option<&Value>) -> Vec<String> {
-    value
-        .and_then(Value::as_array)
-        .map(|values| {
-            values
-                .iter()
-                .filter_map(|value| value.as_str().map(str::trim))
-                .filter(|value| !value.is_empty())
-                .map(str::to_string)
-                .collect::<Vec<_>>()
-        })
-        .unwrap_or_default()
-}
-
-fn artifact_content_bytes(args: &Value) -> Result<Vec<u8>> {
-    if let Some(encoded) = opt_str(args, "content_base64") {
-        return general_purpose::STANDARD
-            .decode(encoded)
-            .map_err(|err| anyhow!("content_base64 is invalid: {err}"));
-    }
-    Ok(args
-        .get("content")
-        .and_then(Value::as_str)
-        .unwrap_or("")
-        .as_bytes()
-        .to_vec())
-}
+// moved artifact_content_bytes -> artifacts.rs
 
 #[derive(Debug, Clone)]
 struct ArtifactUri {
@@ -3419,57 +864,9 @@ impl ArtifactUri {
     }
 }
 
-async fn resolve_artifact_uri(
-    cp: &ControlPlane,
-    current_agent: &str,
-    current_session: &str,
-    artifact_uri: &str,
-    operation: &str,
-) -> Result<(ArtifactUri, crate::gateway::rpc::data_proto::Artifact)> {
-    let uri = parse_artifact_uri(artifact_uri)?;
-    let artifact = cp
-        .kv
-        .get_msg::<crate::gateway::rpc::data_proto::Artifact>(&keys::artifact(
-            &uri.namespace,
-            &uri.agent,
-            &uri.session_id,
-            &uri.artifact_id,
-        ))
-        .await?
-        .ok_or_else(|| anyhow!("Artifact '{}' not found", uri.artifact_id))?;
-    authorize_artifact_access(
-        cp,
-        &uri,
-        current_agent,
-        current_session,
-        operation,
-        artifact_uri,
-    )
-    .await?;
-    Ok((uri, artifact))
-}
+// moved resolve_artifact_uri -> artifacts.rs
 
-fn artifact_json(artifact: &crate::gateway::rpc::data_proto::Artifact) -> Value {
-    let object_ref = artifact.object_ref.as_ref();
-    json!({
-        "id": artifact.id,
-        "sessionId": artifact.session_id,
-        "title": artifact.title,
-        "mediaType": artifact.media_type,
-        "createdByAgent": artifact.created_by_agent,
-        "createdAt": artifact.created_at,
-        "labels": artifact.labels,
-        "metadata": artifact.metadata,
-        "objectRef": object_ref.map(|object| json!({
-            "key": object.key,
-            "mediaType": object.media_type,
-            "sizeBytes": object.size_bytes,
-            "sha256": object.sha256,
-            "filename": object.filename,
-            "metadata": object.metadata,
-        })).unwrap_or_else(|| json!(null))
-    })
-}
+// moved artifact_json -> artifacts.rs
 
 #[derive(Clone, Debug)]
 struct ValidatedHttpUrl {
@@ -3478,375 +875,46 @@ struct ValidatedHttpUrl {
     addrs: Vec<SocketAddr>,
 }
 
-fn http_client(target: &ValidatedHttpUrl) -> Result<reqwest::Client> {
-    let mut builder = reqwest::Client::builder()
-        .timeout(Duration::from_secs(20))
-        .user_agent("TalonResearchBot/0.1 (+https://impalasys.com)")
-        .redirect(reqwest::redirect::Policy::none());
-    if let Some(host) = target.host.as_deref() {
-        builder = builder.resolve_to_addrs(host, &target.addrs);
-    }
-    builder
-        .build()
-        .map_err(|err| anyhow!("research HTTP client failed to build: {err}"))
-}
+// moved http_client -> research.rs
 
-fn validate_http_url(value: &str) -> Result<()> {
-    let url = url::Url::parse(value).map_err(|err| anyhow!("invalid URL: {err}"))?;
-    match url.scheme() {
-        "http" | "https" => Ok(()),
-        scheme => Err(anyhow!("unsupported URL scheme '{}'", scheme)),
-    }
-}
+// moved validate_http_url -> research.rs
 
-async fn validate_public_http_url(value: &str) -> Result<ValidatedHttpUrl> {
-    validate_http_url(value)?;
-    let url = url::Url::parse(value).map_err(|err| anyhow!("invalid URL: {err}"))?;
-    let host = url
-        .host_str()
-        .ok_or_else(|| anyhow!("URL host is required"))?
-        .to_string();
-    let port = url
-        .port_or_known_default()
-        .ok_or_else(|| anyhow!("URL port is required"))?;
+// moved validate_public_http_url -> research.rs
 
-    if let Ok(ip) = host.parse::<IpAddr>() {
-        ensure_public_ip(ip)?;
-        return Ok(ValidatedHttpUrl {
-            url,
-            host: None,
-            addrs: vec![SocketAddr::new(ip, port)],
-        });
-    }
+// moved ensure_public_ip -> research.rs
 
-    let mut addrs = tokio::net::lookup_host((host.clone(), port))
-        .await
-        .map_err(|err| anyhow!("failed to resolve URL host '{host}': {err}"))?;
-    let mut public_addrs = Vec::new();
-    for addr in addrs.by_ref() {
-        ensure_public_ip(addr.ip())?;
-        public_addrs.push(addr);
-    }
-    if public_addrs.is_empty() {
-        return Err(anyhow!("URL host '{host}' resolved no addresses"));
-    }
-    Ok(ValidatedHttpUrl {
-        url,
-        host: Some(host),
-        addrs: public_addrs,
-    })
-}
+// moved is_blocked_ipv4 -> research.rs
 
-fn ensure_public_ip(ip: IpAddr) -> Result<()> {
-    let blocked = match ip {
-        IpAddr::V4(ip) => is_blocked_ipv4(ip),
-        IpAddr::V6(ip) => is_blocked_ipv6(ip),
-    };
-    if blocked {
-        Err(anyhow!("URL resolves to a non-public address"))
-    } else {
-        Ok(())
-    }
-}
+// moved is_blocked_ipv6 -> research.rs
 
-fn is_blocked_ipv4(ip: Ipv4Addr) -> bool {
-    let octets = ip.octets();
-    ip.is_private()
-        || ip.is_loopback()
-        || ip.is_link_local()
-        || ip.is_broadcast()
-        || ip.is_documentation()
-        || ip.is_unspecified()
-        || ip.is_multicast()
-        || octets[0] == 0
-        || octets[0] >= 224
-        || (octets[0] == 100 && (64..=127).contains(&octets[1]))
-        || (octets[0] == 169 && octets[1] == 254)
-}
+// moved extract_title -> research.rs
 
-fn is_blocked_ipv6(ip: Ipv6Addr) -> bool {
-    ip.is_loopback()
-        || ip.is_unspecified()
-        || ip.is_multicast()
-        || (ip.segments()[0] & 0xfe00) == 0xfc00
-        || (ip.segments()[0] & 0xffc0) == 0xfe80
-}
+// moved compact_visible_text -> research.rs
 
-fn extract_title(html: &str) -> String {
-    let lower = html.to_lowercase();
-    let Some(start) = lower.find("<title") else {
-        return String::new();
-    };
-    let Some(open_end) = lower[start..].find('>') else {
-        return String::new();
-    };
-    let content_start = start + open_end + 1;
-    let Some(close) = lower[content_start..].find("</title>") else {
-        return String::new();
-    };
-    decode_html_entities(&html[content_start..content_start + close])
-        .trim()
-        .to_string()
-}
+// moved remove_tag_blocks -> research.rs
 
-fn compact_visible_text(input: &str, max_chars: usize) -> String {
-    let without_scripts = remove_tag_blocks(input, "script");
-    let without_styles = remove_tag_blocks(&without_scripts, "style");
-    let mut text = String::with_capacity(without_styles.len().min(max_chars));
-    let mut in_tag = false;
-    let mut last_was_space = true;
-    for ch in without_styles.chars() {
-        match ch {
-            '<' => {
-                in_tag = true;
-                if !last_was_space {
-                    text.push(' ');
-                    last_was_space = true;
-                }
-            }
-            '>' => in_tag = false,
-            _ if in_tag => {}
-            _ if ch.is_whitespace() => {
-                if !last_was_space {
-                    text.push(' ');
-                    last_was_space = true;
-                }
-            }
-            _ => {
-                text.push(ch);
-                last_was_space = false;
-            }
-        }
-        if text.len() >= max_chars {
-            break;
-        }
-    }
-    decode_html_entities(text.trim())
-}
+// moved decode_html_entities -> research.rs
 
-fn remove_tag_blocks(input: &str, tag: &str) -> String {
-    let mut output = String::with_capacity(input.len());
-    let lower = input.to_lowercase();
-    let open_prefix = format!("<{}", tag);
-    let close = format!("</{}>", tag);
-    let mut pos = 0;
-    while let Some(start_rel) = lower[pos..].find(&open_prefix) {
-        let start = pos + start_rel;
-        output.push_str(&input[pos..start]);
-        if let Some(end_rel) = lower[start..].find(&close) {
-            pos = start + end_rel + close.len();
-        } else {
-            pos = input.len();
-            break;
-        }
-    }
-    output.push_str(&input[pos..]);
-    output
-}
+// moved extract_duckduckgo_results -> research.rs
 
-fn decode_html_entities(input: &str) -> String {
-    input
-        .replace("&amp;", "&")
-        .replace("&quot;", "\"")
-        .replace("&#39;", "'")
-        .replace("&apos;", "'")
-        .replace("&lt;", "<")
-        .replace("&gt;", ">")
-        .replace("&nbsp;", " ")
-}
+// moved extract_attr -> research.rs
 
-fn extract_duckduckgo_results(html: &str, limit: usize) -> Vec<Value> {
-    let mut results: Vec<Value> = Vec::new();
-    let mut pos = 0;
-    while results.len() < limit {
-        let Some(class_rel) = html[pos..].find("result__a") else {
-            break;
-        };
-        let class_pos = pos + class_rel;
-        let anchor_start = html[..class_pos].rfind("<a").unwrap_or(class_pos);
-        let Some(anchor_end_rel) = html[class_pos..].find("</a>") else {
-            break;
-        };
-        let anchor_end = class_pos + anchor_end_rel + "</a>".len();
-        let anchor = &html[anchor_start..anchor_end];
-        let Some(href) = extract_attr(anchor, "href") else {
-            pos = anchor_end;
-            continue;
-        };
-        let Some(url) = normalize_search_result_url(&href) else {
-            pos = anchor_end;
-            continue;
-        };
-        let title = compact_visible_text(anchor, 500);
-        if !title.is_empty() && !results.iter().any(|item| item["url"] == url) {
-            results.push(json!({
-                "title": title,
-                "url": url
-            }));
-        }
-        pos = anchor_end;
-    }
-    results
-}
+// moved normalize_search_result_url -> research.rs
 
-fn extract_attr(input: &str, attr: &str) -> Option<String> {
-    let needle = format!("{}=\"", attr);
-    let start = input.find(&needle)? + needle.len();
-    let end = input[start..].find('"')?;
-    Some(decode_html_entities(&input[start..start + end]))
-}
+// moved default_access_expiry -> artifacts.rs
 
-fn normalize_search_result_url(href: &str) -> Option<String> {
-    if href.starts_with("http://") || href.starts_with("https://") {
-        return Some(href.to_string());
-    }
-    let query_start = href.find("uddg=")? + "uddg=".len();
-    let query = &href[query_start..];
-    let value = query.split('&').next().unwrap_or(query);
-    urlencoding::decode(value)
-        .ok()
-        .map(|value| value.into_owned())
-}
+// moved access_expiry_from_ttl_seconds -> artifacts.rs
 
-fn default_access_expiry() -> i64 {
-    access_expiry_from_ttl_seconds(24 * 60 * 60)
-}
 
-fn access_expiry_from_ttl_seconds(ttl_seconds: i64) -> i64 {
-    if ttl_seconds <= 0 {
-        return default_access_expiry();
-    }
-    let ttl_micros = ttl_seconds.min(MAX_ACCESS_TTL_SECONDS) * 1_000_000;
-    chrono::Utc::now()
-        .timestamp_micros()
-        .saturating_add(ttl_micros)
-}
+// moved parse_artifact_uri -> artifacts.rs
 
-fn normalize_logical_path(path: &str) -> Result<String> {
-    let path = path.trim();
-    if path.is_empty() {
-        return Err(anyhow!("path is required"));
-    }
-    if !path.starts_with('/') {
-        return Err(anyhow!("path must be absolute"));
-    }
-    if path.contains("//") || path.contains('\0') || path.contains("..") {
-        return Err(anyhow!("path is not normalized"));
-    }
-    Ok(path.trim_end_matches('/').to_string())
-}
+// moved validate_uri_segment -> artifacts.rs
 
-fn parse_artifact_uri(uri: &str) -> Result<ArtifactUri> {
-    let rest = uri
-        .trim()
-        .strip_prefix("artifact://")
-        .ok_or_else(|| anyhow!("artifact uri must start with 'artifact://'"))?;
-    let parts = rest.split('/').collect::<Vec<_>>();
-    match parts.as_slice() {
-        [namespace, agent, session_id, artifact_id] => Ok(ArtifactUri {
-            namespace: validate_uri_segment(namespace, "artifact namespace")?,
-            agent: validate_uri_segment(agent, "artifact agent")?,
-            session_id: validate_uri_segment(session_id, "artifact session")?,
-            artifact_id: validate_uri_segment(artifact_id, "artifact id")?,
-        }),
-        _ => Err(anyhow!(
-            "artifact uri must be artifact://<namespace>/<agent>/<session>/<artifact>"
-        )),
-    }
-}
+// moved authorize_artifact_access -> artifacts.rs
 
-fn validate_uri_segment(segment: &str, name: &str) -> Result<String> {
-    if segment.trim().is_empty()
-        || segment.contains('/')
-        || segment.contains('\0')
-        || segment.chars().any(char::is_control)
-    {
-        return Err(anyhow!("{name} segment is invalid"));
-    }
-    Ok(segment.to_string())
-}
 
-async fn authorize_artifact_access(
-    cp: &ControlPlane,
-    uri: &ArtifactUri,
-    current_agent: &str,
-    current_session: &str,
-    operation: &str,
-    artifact_uri: &str,
-) -> Result<()> {
-    if current_agent == uri.agent && current_session == uri.session_id {
-        return Ok(());
-    }
-    if current_agent.trim().is_empty() || current_session.trim().is_empty() {
-        return Err(anyhow!(
-            "artifact uri requires caller agent and session identity"
-        ));
-    }
-    let access = cp
-        .kv
-        .get_msg::<crate::gateway::rpc::data_proto::ArtifactAccess>(&keys::artifact_access(
-            &uri.namespace,
-            &uri.agent,
-            &uri.session_id,
-            &uri.artifact_id,
-            current_agent,
-            current_session,
-        ))
-        .await?
-        .ok_or_else(|| anyhow!("artifact access denied for '{artifact_uri}'"))?;
-    if access.expires_at > 0 && access.expires_at < chrono::Utc::now().timestamp_micros() {
-        return Err(anyhow!("artifact access for '{artifact_uri}' is expired"));
-    }
-    if !access.operations.iter().any(|op| op == operation) {
-        return Err(anyhow!(
-            "artifact access for '{artifact_uri}' does not allow '{operation}'"
-        ));
-    }
-    Ok(())
-}
 
-fn require_capability(spec: &manifests::AgentSpec, capability: &str, action: &str) -> Result<()> {
-    if has_capability_action(spec, capability, action) {
-        return Ok(());
-    }
-    Err(anyhow!(
-        "agent does not have capability '{}:{}'",
-        capability,
-        action
-    ))
-}
 
-fn require_memory_read(spec: &manifests::AgentSpec) -> Result<()> {
-    if has_capability_action(spec, "memory", "read")
-        || has_capability_action(spec, "memory", "inspect")
-    {
-        return Ok(());
-    }
-    Err(anyhow!("agent does not have capability 'memory:read'"))
-}
-
-fn require_file_read(spec: &manifests::AgentSpec) -> Result<()> {
-    if has_capability_action(spec, "files", "read")
-        || has_capability_action(spec, "files", "inspect")
-    {
-        return Ok(());
-    }
-    Err(anyhow!("agent does not have capability 'files:read'"))
-}
-
-fn has_capability_action(spec: &manifests::AgentSpec, capability: &str, action: &str) -> bool {
-    spec.capabilities
-        .get(capability)
-        .map(|actions| {
-            actions.values.iter().any(|value| {
-                matches!(
-                    value.kind.as_ref(),
-                    Some(ProtoValueKind::StringValue(current)) if current == action
-                )
-            })
-        })
-        .unwrap_or(false)
-}
 
 #[cfg(test)]
 mod tests {
@@ -4950,151 +2018,20 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn read_memory_returns_typed_image_output() {
-        let kv = Arc::new(MockKvStore::default());
-        let scheduler = Arc::new(MockScheduler::default());
-        let cp = control_plane(kv.clone(), scheduler);
-        let png_bytes = b"png-bytes";
-        upsert_memory_file(
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "/memory/screenshot.png",
-            "image/png",
-            png_bytes,
-        )
-        .await
-        .unwrap();
-
-        let output = execute_tool_for_session_output(
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "writer",
-            "session-1",
-            &manifests::AgentSpec {
-                capabilities: HashMap::from([(
-                    "memory".to_string(),
-                    crate::gateway::rpc::protobuf_value::ListValue {
-                        values: vec![crate::gateway::rpc::protobuf_value::Value {
-                            kind: Some(ProtoValueKind::StringValue("read".to_string())),
-                        }],
-                    },
-                )]),
-                ..manifests::AgentSpec::default()
-            },
-            READ_MEMORY_TOOL,
-            &json!({"path": "/memory/screenshot.png"}),
-            &Config::default(),
-        )
-        .await
-        .unwrap()
-        .unwrap();
-
-        let source_object_key = output
-            .object_ref()
-            .as_ref()
-            .expect("image output should retain source object ref")
-            .key
-            .clone();
-        let content_parts = output.content_parts();
-        assert_eq!(
-            crate::harness::llm::content_part_object_ref(&content_parts[0])
-                .unwrap()
-                .key,
-            source_object_key
-        );
-
-        let journaled_object = append_test_tool_result(
-            kv.as_ref(),
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "writer",
-            "session-1",
-            READ_MEMORY_TOOL,
-            &output,
-        )
-        .await;
-        assert_eq!(journaled_object.key, source_object_key);
-    }
-
-    #[tokio::test]
-    async fn read_memory_text_output_reuses_object_ref() {
-        let kv = Arc::new(MockKvStore::default());
-        let scheduler = Arc::new(MockScheduler::default());
-        let cp = control_plane(kv.clone(), scheduler);
-        let content = "source text ".repeat(400);
-        upsert_memory_file(
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "/memory/notes.txt",
-            "text/plain; charset=utf-8",
-            content.as_bytes(),
-        )
-        .await
-        .unwrap();
-
-        let output = execute_tool_for_session_output(
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "writer",
-            "session-1",
-            &manifests::AgentSpec {
-                capabilities: HashMap::from([(
-                    "memory".to_string(),
-                    crate::gateway::rpc::protobuf_value::ListValue {
-                        values: vec![crate::gateway::rpc::protobuf_value::Value {
-                            kind: Some(ProtoValueKind::StringValue("read".to_string())),
-                        }],
-                    },
-                )]),
-                ..manifests::AgentSpec::default()
-            },
-            READ_MEMORY_TOOL,
-            &json!({"path": "/memory/notes.txt"}),
-            &Config::default(),
-        )
-        .await
-        .unwrap()
-        .unwrap();
-
-        assert!(output.summary().starts_with("[Object: notes.txt"));
-        assert!(!output.summary().contains(&content));
-        let source_object_key = output
-            .object_ref()
-            .expect("text output should retain source object ref")
-            .key
-            .clone();
-        let content_parts = output.content_parts();
-        assert_eq!(
-            crate::harness::llm::content_part_object_ref(&content_parts[0])
-                .unwrap()
-                .key,
-            source_object_key
-        );
-
-        let journaled_object = append_test_tool_result(
-            kv.as_ref(),
-            &cp,
-            "Tenant:acme:Workspace:main",
-            "writer",
-            "session-1",
-            READ_MEMORY_TOOL,
-            &output,
-        )
-        .await;
-        assert_eq!(journaled_object.key, source_object_key);
-    }
-
-    #[tokio::test]
     async fn read_file_output_reuses_object_ref() {
         let kv = Arc::new(MockKvStore::default());
         let scheduler = Arc::new(MockScheduler::default());
         let cp = control_plane(kv.clone(), scheduler);
         let content = "source text ".repeat(400);
-        let file = upsert_memory_file(
+        let file = upsert_file(
             &cp,
             "Tenant:acme:Workspace:main",
+            None,
             "/memory/notes.txt",
             "text/plain; charset=utf-8",
+            resources_proto::FilePurpose::Artifact as i32,
+            resources_proto::FileIndexPolicy::Search as i32,
+            resources_proto::FileRetention::Retained as i32,
             content.as_bytes(),
         )
         .await
@@ -5123,7 +2060,7 @@ mod tests {
             "Tenant:acme:Workspace:main",
             "writer",
             "session-1",
-            READ_MEMORY_TOOL,
+            READ_FILE_TOOL,
             &output,
         )
         .await;
@@ -6896,3 +3833,5 @@ mod tests {
         assert!(active_context.is_none());
     }
 }
+
+// Facade re-exports for external callers - keep public API stable
